+++ paper/paper.bib
@@ -183,7 +183,8 @@ title = {Systematic and Nonsystematic Search Strategies},
 year = {1992},
 publisher = {Morgan Kaufmann},
 booktitle = {Proceedings of the First International Conference on Artificial Intelligence Planning Systems},
-pages = {145–-152}
+pages = {145–-152},
+doi = {10.1016/b978-0-08-049944-4.50022-7}
 }
 
 @article{Zilberstein1996, 
@@ -203,7 +204,8 @@ pages={73--83}
   booktitle={Proceedings of 1995 IEEE International Conference on Evolutionary Computation}, 
   title={Gaussian mutation and self-adaption for numeric genetic algorithms}, 
   year={1995},
-  pages={384--389}
+  pages={384--389},
+  doi = {10.1109/icec.1995.489178}
 }
 
 @ARTICLE{Szu1987,
@@ -213,7 +215,8 @@ pages={73--83}
   year={1987},
   volume={75},
   number={11},
-  pages={1538--1540}
+  pages={1538--1540},
+  doi = {10.1109/proc.1987.13916}
 }
 
 @inproceedings{Ackley1985,
@@ -231,7 +234,8 @@ title = {Performance of a New Annealing Schedule},
 year = {1988},
 publisher = {IEEE Computer Society Press},
 booktitle = {Proceedings of the 25th ACM/IEEE Design Automation Conference},
-pages = {306–311}
+pages = {306–311},
+doi = {10.1109/dac.1988.14775}
 }
 
 @phdthesis{Swartz1993,
