### 2020-06-30 #2368 @galessiorob
👋 Hi @cicirello!

I am in the process of assigning reviewers to this submission, but if you have any suggestions for reviewers please let me know.

In the meantime, it looks like you're only missing 2 DOIs, if you could nail those down that would great.

Many thanks!
______________

**Basic checks:**

* [x]  There’s a general description of the software
* [x] The software is within scope as research software
* [x] OSI-approved license: GNU General Public License (GPL)
