### 2020-06-30 6731a7f
added a doi and an isbn
