+++ paper/paper.bib
@@ -60,7 +60,8 @@ title = {Genetic Algorithm Parameter Control: Application to Scheduling with Seq
 booktitle = {Proceedings of the 9th International Conference on Bio-inspired Information and Communications Technologies},
 month = {December},
 year = {2015},
-pages = {136--143}
+pages = {136--143},
+doi = {10.4108/eai.3-12-2015.2262346}
 }
 
 @INPROCEEDINGS{Cicirello2014,
@@ -165,7 +166,8 @@ title = {Artificial Intelligence: A Modern Approach},
 year = {2009},
 publisher = {Prentice Hall Press},
 address = {USA},
-edition = {3rd}
+edition = {3rd},
+isbn = {0136042597}
 }
 
 @inproceedings{Bresina1996,
