### 2020-06-30 #2368 @cicirello
Hi @galessiorob,

I added the one actually missing doi.  The other identified by whedon is not correct.  The reference is to a book that doesn't have a doi, and whedon incorrectly identified a doi to someone's review of the book.  I added the isbn of the book instead.  That seems to accomplish essentially the same thing in that the pdf that whedon generated added a link to a permanent record of the book in worldcat.

Here are a couple possible reviewers (mbdemoraes, danieledipompeo, professoralkmin).  I don't know them, and don't know their availability.  I found them by skimming JOSS's reviewer list.
