+++ paper/paper.bib
@@ -138,6 +138,7 @@ doi = {10.1007/s10732-005-6997-8}
   title     = {Computers and Intractability: {A} Guide to the Theory of NP-Completeness},
   publisher = {W. H. Freeman},
   year      = {1979},
+  isbn = {0716710455}
 }
 
 @book{Hoos2005,
@@ -145,7 +146,8 @@ doi = {10.1007/s10732-005-6997-8}
                Thomas St{\"{u}}tzle},
   title     = {Stochastic Local Search: Foundations and Applications},
   publisher = {Morgan Kaufmann},
-  year      = {2005}
+  year      = {2005},
+  isbn = {1558608729}
 }
 
 @article {Kirkpatrick1983,
