### 2020-07-08 #2448 @whedon
[REVIEW]: Chips-n-Salsa: A Java Library of Customizable, Hybridizable, Iterative, Parallel, Stochastic, and Self-Adaptive Local Search Algorithms
**Submitting author:** @cicirello (<a href="http://orcid.org/0000-0003-1072-8559">Vincent A Cicirello</a>)
**Repository:** <a href="https://github.com/cicirello/Chips-n-Salsa" target ="_blank">https://github.com/cicirello/Chips-n-Salsa</a>
**Version:** v1.3.0
