### 2020-07-08 7d65640
Constructive heuristics, and related
Added code for generating solutions to permutation optimization problems
via constructive heuristics. This is in preparation for adding heuristic
stochastic sampling algorithms (HBSS and VBSS) which use constructive
heuristics.
