+++ tests/org/cicirello/search/ss/VBSSTests.java
@@ -1,0 +1,611 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+ 
+package org.cicirello.search.ss;
+
+
+import org.junit.*;
+import static org.junit.Assert.*;
+import org.cicirello.permutations.Permutation;
+import org.cicirello.search.problems.Problem;
+import org.cicirello.search.problems.IntegerCostOptimizationProblem;
+import org.cicirello.search.problems.OptimizationProblem;
+import org.cicirello.search.SolutionCostPair;
+import org.cicirello.search.ProgressTracker;
+
+/**
+ * JUnit tests for ValueBiasedStochasticSampling.
+ */
+public class VBSSTests {
+	
+	@Test
+	public void testWithIntCosts() {
+		for (int n = 0; n < 10; n++) {
+			IntProblem problem = new IntProblem();
+			IntHeuristic h = new IntHeuristic(problem, n);
+			ValueBiasedStochasticSampling ch = new ValueBiasedStochasticSampling(h);
+			assertEquals(0, ch.getTotalRunLength());
+			assertTrue(problem == ch.getProblem());
+			ProgressTracker<Permutation> tracker = ch.getProgressTracker();
+			SolutionCostPair<Permutation> solution = ch.optimize();
+			assertEquals(1, ch.getTotalRunLength());
+			assertEquals((n+1)*n/2, solution.getCost());
+			assertEquals((n+1)*n/2, tracker.getCost());
+			Permutation p = solution.getSolution();
+			assertEquals(n, p.length());
+			solution = ch.optimize();
+			assertEquals(2, ch.getTotalRunLength());
+			assertEquals((n+1)*n/2, solution.getCost());
+			assertEquals((n+1)*n/2, tracker.getCost());
+			tracker = new ProgressTracker<Permutation>();
+			ch.setProgressTracker(tracker);
+			assertTrue(tracker == ch.getProgressTracker());
+		}
+	}
+	
+	@Test
+	public void testWithDoubleCosts() {
+		for (int n = 0; n < 10; n++) {
+			DoubleProblem problem = new DoubleProblem();
+			DoubleHeuristic h = new DoubleHeuristic(problem, n);
+			ValueBiasedStochasticSampling ch = new ValueBiasedStochasticSampling(h);
+			assertEquals(0, ch.getTotalRunLength());
+			assertTrue(problem == ch.getProblem());
+			ProgressTracker<Permutation> tracker = ch.getProgressTracker();
+			SolutionCostPair<Permutation> solution = ch.optimize();
+			assertEquals(1, ch.getTotalRunLength());
[…]
