+++ src/org/cicirello/search/ss/ValueBiasedStochasticSampling.java
@@ -1,0 +1,482 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.cicirello.search.ss;
+
+import org.cicirello.permutations.Permutation;
+import org.cicirello.search.SimpleMetaheuristic;
+import org.cicirello.search.Metaheuristic;
+import org.cicirello.search.problems.Problem;
+import org.cicirello.search.problems.IntegerCostOptimizationProblem;
+import org.cicirello.search.SolutionCostPair;
+import org.cicirello.search.ProgressTracker;
+import java.util.concurrent.ThreadLocalRandom;
+
+/**
+ * <p>Value Biased Stochastic Sampling (VBSS) is a form of stochastic sampling 
+ * search that uses a constructive heuristic to bias the random decisions.
+ * In VBSS, the search generates N random candidate solutions to the 
+ * problem, using a problem-specific heuristic to bias each random decision in favor
+ * of choices that are preferred by the heuristic.
+ * It evaluates each of the N candidate solutions with respect to the optimization
+ * problem's cost function, and returns the best of the N candidate solutions.</p>
+ *
+ * <p>Although the VBSS algorithm itself is not technically restricted to 
+ * permutation problems, the ValueBiasedStochasticSampling class only supports
+ * optimization problems over the space of permutations.</p>
+ *
+ * <p>Let's consider an illustrative example.  Consider a problem whose solution
+ * is to be represented with a permutation of length L of the integers 
+ * {0, 1, ..., (L-1)}.  For the sake of this example, let L=4.  Thus, we are searching
+ * for a permutation of the integers {0, 1, 2, 3}.  We will iteratively build up
+ * a partial permutation containing a subset of the elements into a complete
+ * permutation.</p>  
+ *
+ * <p>We begin with an empty partial permutation: p = [].  The first iteration
+ * will select the first element.  To do so, we use a heuristic to evaluate each option
+ * within the context of the problem.  Let S be the set of elements not yet in the 
+ * permutation, which in this case is initially S = {0, 1, 2, 3}.  Let h(p, e) 
+ * be a constructive
+ * heuristic that takes as input the current partial permutation p, and an element e from S
+ * under consideration for addition to p, and which produces a real value as output 
+ * that increases as the importance of adding e to p increases.  That is, higher values
+ * of the heuristic imply that the heuristic has a higher level of confidence that the
+ * element e should be added next to p.  For the sake of this example, consider that the
[…]
