### 2020-07-09 f2c8d64
Minor edits to JOSS paper
Minor edits to JOSS paper to reflect some additional algorithms added to
library.
