+++ paper/paper.bib
@@ -132,6 +132,15 @@ PAGES = {5--34},
 doi = {10.1007/s10732-005-6997-8}
 }
 
+@PHDTHESIS{Cicirello2003,
+AUTHOR = {Vincent A. Cicirello},
+TITLE = {Boosting Stochastic Problem Solvers Through Online Self-Analysis of Performance},
+SCHOOL = {The Robotics Institute, School of Computer Science, Carnegie Mellon University},
+YEAR = {2003},
+ADDRESS = {Pittsburgh, PA},
+MONTH = {July}
+}
+
 @book{Garey1979,
   author    = {Michael R. Garey and
                David S. Johnson},
