+++ paper/paper.md
@@ -29,7 +29,7 @@ The repository (https://github.com/cicirello/Chips-n-Salsa) contains source code
 
 # Chips-n-Salsa Features and Functionality:
 
-* __Stochastic Local Search Algorithms__: The library most extensively supports simulated annealing [@Kirkpatrick1983], providing implementations of all of the common annealing schedules, as well as more advanced annealing schedules.  However, it also includes implementations of common forms of hill climbing [@Rusell2009], as well as iterative sampling [@Langley1992]. Additional algorithms from prior research will be integrated with the library in the near future, such as HBSS [@Bresina1996] and VBSS [@Cicirello2005]. We carefully chose the random number generators for the library based on our prior research [@Cicirello2018a] to minimize the runtime impact from one of the more costly operations of a metaheuristic.
+* __Stochastic Local Search Algorithms__: The library extensively supports simulated annealing [@Kirkpatrick1983], providing implementations of all of the common annealing schedules, as well as more advanced annealing schedules.  It also includes implementations of common forms of hill climbing [@Rusell2009], as well as several stochastic sampling search algorithms, such as iterative sampling [@Langley1992], Heuristic Biased Stochastic Sampling (HBSS) [@Bresina1996], and Value Biased Stochastic Sampling (VBSS) [@Cicirello2005; @Cicirello2003]. We carefully chose the random number generators for the library based on our prior research [@Cicirello2018a] to minimize the runtime impact from one of the more costly operations of a metaheuristic.
 * __Customizable__: Chips-n-Salsa extensively uses Java's generic types enabling using the library to optimize other types of representations beyond what is provided in the library. It is also designed to allow easily integrating new mutation operators, as well as to enable customizing all stages of the local search (e.g., choice of annealing schedule for simulated annealing).
 * __Hybridizable__: Chips-n-Salsa supports integrating multiple forms of local search (e.g., hybrids of hill climbing with simulated annealing), creating hybrid mutation operators (e.g., combining multiple mutation operators), as well as support for running more than one type of search for the same problem concurrently using multiple threads as a form of algorithm portfolio [@Gomes2001].
 * __Iterative__: Chips-n-Salsa supports multistart metaheuristics, including implementations of several restart schedules [@Luby1993; @Cicirello2017] for varying the run lengths across the restarts. 
