### 2020-07-09 f148295
support for exponential bias
Added support for exponential bias to HBSS implementation.
