### 2020-07-10 ace2b0a
moved increment of numGenerated
