+++ src/org/cicirello/search/ss/HeuristicBiasedStochasticSampling.java
@@ -299,6 +299,7 @@ public final class HeuristicBiasedStochasticSampling implements SimpleMetaheuris
 	@Override
 	public SolutionCostPair<Permutation> optimize() {
 		if (tracker.didFindBest() || tracker.isStopped()) return null;
+		numGenerated++;
 		return sampler.optimize();
 	}
 	
@@ -321,6 +322,7 @@ public final class HeuristicBiasedStochasticSampling implements SimpleMetaheuris
 		SolutionCostPair<Permutation> best = null;
 		for (int i = 0; i < numSamples && !tracker.didFindBest() && !tracker.isStopped(); i++) {
 			SolutionCostPair<Permutation> current = sampler.optimize();
+			numGenerated++;
 			if (best == null || current.compareTo(best) < 0) best = current;
 		}
 		return best;
@@ -509,7 +511,6 @@ public final class HeuristicBiasedStochasticSampling implements SimpleMetaheuris
 					p.extend(which);
 				}
 			}
-			numGenerated++;
 			int cost = incEval.cost();
 			Permutation complete = p.toComplete();
 			if (cost < tracker.getCost()) {
@@ -569,7 +570,6 @@ public final class HeuristicBiasedStochasticSampling implements SimpleMetaheuris
 					p.extend(which);
 				}
 			}
-			numGenerated++;
 			double cost = incEval.cost();
 			Permutation complete = p.toComplete();
 			if (cost < tracker.getCostDouble()) {
