### 2020-07-10 9e7cb76
Added acceptance band sampling
