+++ src/org/cicirello/search/ss/AcceptanceBandSampling.java
@@ -1,0 +1,401 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.cicirello.search.ss;
+
+import org.cicirello.permutations.Permutation;
+import org.cicirello.search.SimpleMetaheuristic;
+import org.cicirello.search.Metaheuristic;
+import org.cicirello.search.problems.Problem;
+import org.cicirello.search.problems.IntegerCostOptimizationProblem;
+import org.cicirello.search.SolutionCostPair;
+import org.cicirello.search.ProgressTracker;
+import org.cicirello.math.rand.RandomIndexer;
+
+/**
+ * <p>The AcceptanceBandSampling class implements a form of stochastic sampling 
+ * search that uses a constructive heuristic to guide the random decisions.
+ * When making a random decision, all options are evaluated with the heuristic.
+ * An acceptance band is then defined based upon the option with the highest
+ * heuristic evaluation.  Specifically, all options with a heuristic evaluation 
+ * within B% of the highest heuristic evaluation of the available options are
+ * considered equivalent.  A choice is then made uniformly at random from the
+ * set of equivalents.</p>
+ * 
+ * <p>The search generates N random candidate solutions to the 
+ * problem, using a problem-specific heuristic for guidance.
+ * It evaluates each of the N candidate solutions with respect to the optimization
+ * problem's cost function, and returns the best of the N candidate solutions.</p>
+ *
+ * <p>Although the algorithm itself is not technically restricted to 
+ * permutation problems, the AcceptanceBandSampling class only supports
+ * optimization problems over the space of permutations.</p>
+ *
+ * <p>The acceptance bands are defined in terms of a parameter &beta;,
+ * which must be in the interval [0.0, 1.0].  Imagine that we have a set of k
+ * alternatives, a[0], a[1], ..., a[k], to pick from that have heuristic 
+ * values: h[0], h[1], ..., h[k].
+ * We assume in this implementation that higher heuristic values imply the
+ * options perceived better by the heuristic.  We compute:
+ * h' = max {h[0], h[1], ..., h[k]}.  We define an acceptance threshold:
+ * T = (1.0 - beta)h'.  The set S of equivalent choices is then
+ * computed as: S = { a[k] | h[k] &ge; T}.  We then choose uniformly at
+ * random from the set S.</p>
+ *
+ * <p>If beta=1.0, then all alternatives are considered equivalent (we assume
+ * heuristic values are non-negative) since the threshold T would be 0.0
+ * regardless of the heuristic values.  If beta=0.0, then only the choices
[…]
