+++ paper/paper.bib
@@ -289,4 +289,25 @@ doi = {10.1016/S0004-3702(00)00081-3},
 author = {Carla P. Gomes and Bart Selman}
 }
 
+@inproceedings{Oddi1997,
+author = {Oddi, Angelo and Smith, Stephen F.},
+title = {Stochastic Procedures for Generating Feasible Schedules},
+year = {1997},
+publisher = {AAAI Press},
+booktitle = {Proceedings of the Fourteenth National Conference on Artificial Intelligence},
+pages = {308–-314}
+}
+
+@inproceedings{Gomes1998,
+author = {Gomes, Carla P. and Selman, Bart and Kautz, Henry},
+title = {Boosting Combinatorial Search through Randomization},
+year = {1998},
+publisher = {AAAI Press},
+booktitle = {Proceedings of the Fifteenth National Conference on Artificial Intelligence},
+pages = {431–-437}
+}
+
+
+
+
 
