### 2020-07-13 c02bf0c
refactored constructive heuristic related interfaces
Previous interfaces ConstructiveHeuristc and IncermentalEvaluation were
more complex than necessary.  Simplified these.
