+++ src/org/cicirello/search/ss/ValueBiasedStochasticSampling.java
@@ -157,7 +157,7 @@ public final class ValueBiasedStochasticSampling implements SimpleMetaheuristic<
 	 * @param heuristic The constructive heuristic.
 	 * @throws NullPointerException if heuristic is null
 	 */
-	public ValueBiasedStochasticSampling(ConstructiveHeuristic<? extends Number> heuristic) {
+	public ValueBiasedStochasticSampling(ConstructiveHeuristic heuristic) {
 		this(heuristic, null, new ProgressTracker<Permutation>());
 	}
 	
@@ -170,7 +170,7 @@ public final class ValueBiasedStochasticSampling implements SimpleMetaheuristic<
 	 * @param tracker A ProgressTracker
 	 * @throws NullPointerException if heuristic or tracker is null
 	 */
-	public ValueBiasedStochasticSampling(ConstructiveHeuristic<? extends Number> heuristic, ProgressTracker<Permutation> tracker) {
+	public ValueBiasedStochasticSampling(ConstructiveHeuristic heuristic, ProgressTracker<Permutation> tracker) {
 		this(heuristic, null, tracker);
 	}
 	
@@ -181,7 +181,7 @@ public final class ValueBiasedStochasticSampling implements SimpleMetaheuristic<
 	 * @param exponent The bias function is defined as: bias(value) = pow(value, exponent).
 	 * @throws NullPointerException if heuristic is null
 	 */
-	public ValueBiasedStochasticSampling(ConstructiveHeuristic<? extends Number> heuristic, double exponent) {
+	public ValueBiasedStochasticSampling(ConstructiveHeuristic heuristic, double exponent) {
 		this(heuristic, exponent, new ProgressTracker<Permutation>());
 	}
 	
@@ -192,7 +192,7 @@ public final class ValueBiasedStochasticSampling implements SimpleMetaheuristic<
 	 * @param tracker A ProgressTracker
 	 * @throws NullPointerException if heuristic or tracker is null
 	 */
-	public ValueBiasedStochasticSampling(ConstructiveHeuristic<? extends Number> heuristic, double exponent, ProgressTracker<Permutation> tracker) {
+	public ValueBiasedStochasticSampling(ConstructiveHeuristic heuristic, double exponent, ProgressTracker<Permutation> tracker) {
 		this(heuristic, 
 			new BiasFunction() { 
 				public double bias(double value) { 
@@ -210,7 +210,7 @@ public final class ValueBiasedStochasticSampling implements SimpleMetaheuristic<
 	 * @param bias The bias function.  If null, then the default bias is used.
 	 * @throws NullPointerException if heuristic is null
 	 */
-	public ValueBiasedStochasticSampling(ConstructiveHeuristic<? extends Number> heuristic, BiasFunction bias) {
+	public ValueBiasedStochasticSampling(ConstructiveHeuristic heuristic, BiasFunction bias) {
 		this(heuristic, bias, new ProgressTracker<Permutation>());
 	}
 	
@@ -221,7 +221,7 @@ public final class ValueBiasedStochasticSampling implements SimpleMetaheuristic<
 	 * @param tracker A ProgressTracker
 	 * @throws NullPointerException if heuristic or tracker is null
 	 */
-	public ValueBiasedStochasticSampling(ConstructiveHeuristic<? extends Number> heuristic, BiasFunction bias, ProgressTracker<Permutation> tracker) {
+	public ValueBiasedStochasticSampling(ConstructiveHeuristic heuristic, BiasFunction bias, ProgressTracker<Permutation> tracker) {
 		if (heuristic == null || tracker == null) {
 			throw new NullPointerException();
 		}
[…]
