+++ tests/org/cicirello/search/ss/HBSSTests.java
@@ -599,7 +599,7 @@ public class HBSSTests {
 	 * designed to prefer even permutation elements (largest to smallest), followed by odd
 	 * (largest to smallest).
 	 */
-	private static class IntHeuristic implements ConstructiveHeuristic<Integer> {
+	private static class IntHeuristic implements ConstructiveHeuristic {
 		private IntProblem problem;
 		private int n;
 		public IntHeuristic(IntProblem problem, int n) { this.problem = problem; this.n = n; }
@@ -608,7 +608,7 @@ public class HBSSTests {
 		@Override public IntIncEval createIncrementalEvaluation() {
 			return new IntIncEval();
 		}
-		@Override public double h(PartialPermutation p, int element, IncrementalEvaluation<Integer> incEval) {
+		@Override public double h(PartialPermutation p, int element, IncrementalEvaluation incEval) {
 			IntIncEval inc = (IntIncEval)incEval;
 			if (element % 2 == 0) return n + element;
 			else return element;
@@ -620,7 +620,7 @@ public class HBSSTests {
 	 * designed to prefer even permutation elements (largest to smallest), followed by odd
 	 * (largest to smallest).
 	 */
-	private static class DoubleHeuristic implements ConstructiveHeuristic<Double> {
+	private static class DoubleHeuristic implements ConstructiveHeuristic {
 		private DoubleProblem problem;
 		private int n;
 		public DoubleHeuristic(DoubleProblem problem, int n) { this.problem = problem; this.n = n; }
@@ -629,7 +629,7 @@ public class HBSSTests {
 		@Override public DoubleIncEval createIncrementalEvaluation() {
 			return new DoubleIncEval();
 		}
-		@Override public double h(PartialPermutation p, int element, IncrementalEvaluation<Double> incEval) {
+		@Override public double h(PartialPermutation p, int element, IncrementalEvaluation incEval) {
 			DoubleIncEval inc = (DoubleIncEval)incEval;
 			if (element % 2 == 0) return n + element;
 			else return element;
@@ -639,18 +639,16 @@ public class HBSSTests {
 	/*
 	 * Fake designed for predictable test cases.
 	 */
-	private static class IntIncEval implements IncrementalEvaluation<Integer> {
+	private static class IntIncEval implements IncrementalEvaluation {
 		private int sum;
-		@Override public Integer cost() { return sum; }
 		@Override public void extend(PartialPermutation p, int element) { sum += element + 1; }
 	}
 	
 	/*
 	 * Fake designed for predictable test cases.
 	 */
-	private static class DoubleIncEval implements IncrementalEvaluation<Double> {
+	private static class DoubleIncEval implements IncrementalEvaluation {
 		private int sum;
-		@Override public Double cost() { return 1.0*sum; }
 		@Override public void extend(PartialPermutation p, int element) { sum += element + 1; }
 	}
 	
@@ -659,7 +657,13 @@ public class HBSSTests {
 	 * Fake problem. Doesn't really matter for what we are testing.
 	 */
 	private static class IntProblem implements IntegerCostOptimizationProblem<Permutation> {
-		@Override public int cost(Permutation candidate) { return 5; }
+		@Override public int cost(Permutation candidate) { 
+			int sum = 0;
+			for (int i = 0; i < candidate.length(); i++) {
+				sum += candidate.get(i);
+			}
+			return sum + candidate.length(); 
+		}
 		@Override public int value(Permutation candidate) { return cost(candidate); }
 	}
 	
@@ -668,7 +672,13 @@ public class HBSSTests {
 	 * Fake problem. Doesn't really matter for what we are testing.
 	 */
 	private static class DoubleProblem implements OptimizationProblem<Permutation> {
[…]

[…]
