### 2020-07-13 44603d1
added single machine scheduling interface
