+++ src/org/cicirello/search/problems/scheduling/SingleMachineScheduling.java
@@ -1,0 +1,221 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.cicirello.search.problems.scheduling;
+
+import org.cicirello.permutations.Permutation;
+
+/**
+ * <p>Classes that implement single machine scheduling problems should
+ * implement this interface to enable heuristic, etc to interact with the
+ * data that defines the jobs of the scheduling instance, such as the
+ * process times, setup times (if any), due-dates, etc of the jobs.</p>
+ *
+ * <p>This interface includes methods for accessing all of the common
+ * elements of scheduling problems: processing times, due dates, release
+ * dates, weights, setup times.  It also has methods for checking whether
+ * each of these (except processing times) are present.  Some scheduling
+ * cost functions are influenced only by some of these, etc.  Default
+ * implementations are included for all of these potential properties
+ * except for processing times, where the defaults make the obvious
+ * interpretation if that property is not present.  For example,
+ * the default release date is 0, for available for scheduling right
+ * from the start of the schedule, which is the assumption made by many
+ * scheduling problems.  See the various method documentation for
+ * details of defaults for the others.</p>
+ *
+ * <p>The design rationale for providing access to all of these 
+ * properties, with defaults, is to have a generalized interface to
+ * scheduling problem data to enable easily mixing and matching
+ * implementations of scheduling instance generators with 
+ * scheduling cost functions.  For example, with this approach,
+ * we can take a scheduling problem generator for a common due date
+ * problem (which has both early and late weights) and instead optimize
+ * weighted tardiness (which doesn't care about early weights).
+ * Therefore, we have separated the scheduling cost functions from
+ * the classes that represent scheduling problem characteristics
+ * in the library.</p>
+ *
+ * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
+ * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
+ * @version 7.13.2020
+ */
+public interface SingleMachineScheduling {
+	
+	/**
+	 * Gets the number of jobs of this scheduling problem instance.
+	 * @return the number of jobs
+	 */
+	int numberOfJobs();
+	
+	/**
+	 * Gets the processing time of a job, which is the amount of
+	 * time required by the machine to process the job.
+	 * @param j The index of the job, which must be in the interval: [0, numberOfJobs()).
+	 * @return the processing time of job j.
+	 * @throws IndexOutOfBoundsException if j &lt; 0 or j &ge; numberOfJobs() 
+	 */
+	int getProcessingTime(int j);
+	
+	/**
+	 * Computes the completion times of all of the jobs if they were scheduled in
[…]
