### 2020-07-14 a2b8faf
single machine scheduling update
Renamed interface. Added single machine scheduling with setups problem
instance generation.
