+++ tests/org/cicirello/search/problems/scheduling/WeightedStaticSetupsTests.java
@@ -1,0 +1,570 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.cicirello.search.problems.scheduling;
+
+import org.junit.*;
+import static org.junit.Assert.*;
+import org.cicirello.permutations.Permutation;
+
+/**
+ * JUnit tests for the WeightedStaticSchedulingWithSetups class.
+ */
+public class WeightedStaticSetupsTests {
+	
+	@Test
+	public void testCorrectNumJobs() {
+		double[] tau = {0.0, 0.25, 0.5, 0.75, 1.0};
+		double[] r = {0.0, 0.25, 0.5, 0.75, 1.0};
+		double[] eta = {0.0, 0.25, 0.5, 0.75, 1.0};
+		for (int n = 1; n <= 5; n++) {
+			for (int i = 0; i < tau.length; i++) {
+				for (int j = 0; j < r.length; j++) {
+					for (int k = 0; k < eta.length; k++) {
+						WeightedStaticSchedulingWithSetups s = new WeightedStaticSchedulingWithSetups(n, tau[i], r[j], eta[k], 42);
+						assertEquals(n, s.numberOfJobs());
+						assertTrue(s.hasDueDates());
+						assertTrue(s.hasWeights());
+						assertTrue(s.hasSetupTimes());
+						assertFalse(s.hasEarlyWeights());
+						assertFalse(s.hasReleaseDates());
+					}
+				}
+			}
+		}
+	}
+	
+	@Test
+	public void testCompletionTimeCalculation() {
+		int n = 5;
+		double[] tau = {0.0, 0.25, 0.5, 0.75, 1.0};
+		double[] r = {0.0, 0.25, 0.5, 0.75, 1.0};
+		double[] eta = {0.0, 0.25, 0.5, 0.75, 1.0};
+		for (int i = 0; i < tau.length; i++) {
+			for (int j = 0; j < r.length; j++) {
+				for (int k = 0; k < eta.length; k++) {
+					WeightedStaticSchedulingWithSetups s = new WeightedStaticSchedulingWithSetups(n, tau[i], r[j], eta[k], 42);
+					Permutation p1 = new Permutation(new int[] {0, 1, 2, 3, 4});
+					int[] c1 = s.getCompletionTimes(p1);
+					int expected = 0;
+					int last = n;
+					for (int x = 0; x < n; x++) {
+						expected += s.getProcessingTime(x) + s.getSetupTime(last,x);
+						assertEquals(expected, c1[x]);
+						last=x;
+					}
[…]
