+++ src/org/cicirello/search/problems/scheduling/WeightedStaticSchedulingWithSetups.java
@@ -1,0 +1,276 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.cicirello.search.problems.scheduling;
+
+import org.cicirello.permutations.Permutation;
+import org.cicirello.math.rand.RandomIndexer;
+import java.util.SplittableRandom;
+
+/**
+ * <p>This class provides a representation of, and means of generating,
+ * instances of single machine scheduling problems involving weights,
+ * due dates, and sequence-dependent setup times, but without release
+ * dates (i.e., all jobs are released at the start of the problem at time 0,
+ * thus, the term "static" in the class name).</p>
+ *
+ * <p>This class generates instances using a procedure largely based upon
+ * that which was used to generate the benchmark instances for 
+ * weighted tardiness scheduling with sequence-dependent setups described
+ * in the following paper, and with instances 
+ * now available at the following Harvard Dataverse site:</p>
+ * <ul>
+ * <li>Vincent A. Cicirello. 
+ * <a href="https://www.cicirello.org/publications/cicirello2003cmu.html">Weighted 
+ * Tardiness Scheduling with Sequence-Dependent Setups: A Benchmark Library</a>.
+ * Technical Report, Intelligent Coordination and Logistics Laboratory, Robotics 
+ * Institute, Carnegie Mellon University, Pittsburgh, PA, February 2003.</li>
+ * <li>Vincent A. Cicirello.
+ * <a href="http://dx.doi.org/10.7910/DVN/VHA0VQ">Weighted 
+ * Tardiness Scheduling with Sequence-Dependent Setups: A Benchmark Library</a>.
+ * Harvard Dataverse, doi:10.7910/DVN/VHA0VQ, June 2016.</li>
+ * </ul>
+ * <p>That approach was based on the description of generating instances 
+ * for that scheduling problem described in the following paper:</p>
+ * <ul>
+ * <li>Lee, Y. H., Bhaskaran, K., and Pinedo, M. A heuristic to minimize the 
+ * total weighted tardiness with sequence-dependent setups. IIE Transactions, 
+ * 29:45–52, 1997.</li>
+ * </ul>
+ * <p>The Chips-n-Salsa library separates the representation of the scheduling
+ * problem instance data (e.g., processing times, setup times, etc) from the
+ * implementations of scheduling cost functions deliberately to enable
+ * defining additional problems (i.e., different cost functions to optimize)
+ * using the same scheduling problem generators.  This is very much like
[…]
