### 2020-07-15 cicirello/Chips-n-Salsa#2 @nnadeau
Simplify entrypoint
When getting started with the repo, the two primary entrypoints are "building the code" and "running examples".

While a java developer will have no problem building with `ant` and understanding why the `-cp` flag is needed to run the examples, these are an extra impedance for someone new to the repo.

I'd recommend using a `Makefile` as the entrypoint. See https://github.com/cicirello/Chips-n-Salsa/pull/1 for an example. The READMEs will need to be updated as well.

Moreover, the two READMEs (README.md and examples/README.md) cite the same way to run the examples. This duplication of instructions can lead to future stale or out-of-sync documentation. Using a Makefile separates the implementation from the documentation.

cc https://github.com/openjournals/joss-reviews/issues/2448
