### 2020-07-15 cicirello/Chips-n-Salsa#3 @nnadeau
Specifiy Java version
The supported java (and openjdk) versions should be explicitly specified in the README.

cc https://github.com/openjournals/joss-reviews/issues/2448
