### 2020-07-15 cicirello/Chips-n-Salsa#4 @nnadeau
Set up automated testing and CI/CD
Good software practice calls for automated testing, continuous integrations, and continuous deployments. Each pull request, change, merge, and release should be rigorously tested and deployed without human intervention (the source of all errors).

Set up testing/CI/CD with one of the many free services, such as GitHub Actions, CircleCI, TravisCI, etc

cc https://github.com/openjournals/joss-reviews/issues/2448
