### 2020-07-15 cicirello/Chips-n-Salsa#5 @nnadeau
References and Citations
- There are 27 citations for a two page paper. Are all these citations needed? Are they all truly relevant?
- 11 of the 27 citations (40%) are self-citations. This can be seen as citation farming. Same questions as above.
- Many of the non-self-citations are >5 years old. Due to the popularity of AI/ML, the fields of optimization and search algorithms are rapidly evolving. Are these the most up-to-date and relevant references?

cc https://github.com/openjournals/joss-reviews/issues/2448
