### 2020-07-15 cicirello/Chips-n-Salsa#6 @nnadeau
Other libraries?
The paper includes a lot of discussions around algorithms (the first section) and the library itself (the second section). What about other (i.e., competitor) libraries? I highly doubt that this is the only library that serves this purpose for Java. What about for other languages? How does this work compare to those implementations?

cc https://github.com/openjournals/joss-reviews/issues/2448
