### 2020-07-15 cicirello/Chips-n-Salsa#7 @nnadeau
Installation, usage, and artifacts
The README explains how to build the repo and run the examples, but not how (and why) to integrate it into other codebases. There should also be an artifactory or release management system (e.g., JFrog, Nexus, Maven Central Repository, GitHub releases) that contains the latest built JARs for quick integration. Bonus points if the code can be integrated through the most popular Java build tools, such as `gradle` and `maven`.

cc https://github.com/openjournals/joss-reviews/issues/2448
