### 2020-07-15 cicirello/Chips-n-Salsa#8 @nnadeau
Copyright printing should be in its own file/class/utility package to avoid code duplication
https://github.com/cicirello/Chips-n-Salsa/blob/afc59218f79248b2b079389172b491fcaf013fc2/examples/org/cicirello/examples/chipsnsalsa/CustomIntegerCostProblemExample.java#L41

cc https://github.com/openjournals/joss-reviews/issues/2448
