+++ tests/org/cicirello/search/problems/scheduling/WeightedStaticSetupsTests.java
@@ -86,70 +86,71 @@ public class WeightedStaticSetupsTests {
 	
 	@Test
 	public void testParamsNon0Non1() {
-		int n = 10;
-		double[] tau = {0.25, 0.5, 0.75};
-		double[] r = {0.25, 0.5, 0.75};
-		double[] eta = {0.25, 0.5, 0.75};
-		for (int i = 0; i < tau.length; i++) {
-			for (int j = 0; j < r.length; j++) {
-				for (int k = 0; k < eta.length; k++) {
-					WeightedStaticSchedulingWithSetups s = new WeightedStaticSchedulingWithSetups(n, tau[i], r[j], eta[k], 42);
-					assertEquals(n, s.numberOfJobs());
-					assertTrue(s.hasDueDates());
-					assertTrue(s.hasWeights());
-					assertTrue(s.hasSetupTimes());
-					assertFalse(s.hasEarlyWeights());
-					assertFalse(s.hasReleaseDates());
-					boolean diffP = false;
-					boolean diffW = false;
-					boolean diffS0 = false;
-					boolean diffD = false;
-					int pSum = 0;
-					int sSum = 0;
-					for (int x = 0; x < n; x++) {
-						pSum += s.getProcessingTime(x);
-						boolean diffS = false;
-						for (int y = 0; y <= n; y++) {
-							sSum += s.getSetupTime(y,x);
-							if (y > 0 && !diffS && s.getSetupTime(y,x) != s.getSetupTime(y-1,x)) {
-								diffS = true;
+		for (int n = 8; n <= 512; n *= 8) {
+			double[] tau = {0.25, 0.5, 0.75};
+			double[] r = {0.25, 0.5, 0.75};
+			double[] eta = {0.25, 0.5, 0.75};
+			for (int i = 0; i < tau.length; i++) {
+				for (int j = 0; j < r.length; j++) {
+					for (int k = 0; k < eta.length; k++) {
+						WeightedStaticSchedulingWithSetups s = new WeightedStaticSchedulingWithSetups(n, tau[i], r[j], eta[k], 42);
+						assertEquals(n, s.numberOfJobs());
+						assertTrue(s.hasDueDates());
+						assertTrue(s.hasWeights());
+						assertTrue(s.hasSetupTimes());
+						assertFalse(s.hasEarlyWeights());
+						assertFalse(s.hasReleaseDates());
+						boolean diffP = false;
+						boolean diffW = false;
+						boolean diffS0 = false;
+						boolean diffD = false;
+						int pSum = 0;
+						int sSum = 0;
+						for (int x = 0; x < n; x++) {
+							pSum += s.getProcessingTime(x);
+							boolean diffS = false;
+							for (int y = 0; y <= n; y++) {
+								sSum += s.getSetupTime(y,x);
+								if (y > 0 && !diffS && s.getSetupTime(y,x) != s.getSetupTime(y-1,x)) {
+									diffS = true;
+								}
+							}
+							assertTrue(diffS);
+							if (x > 0 && !diffP && s.getProcessingTime(x) != s.getProcessingTime(x-1)) {
+								diffP = true;
+							}
+							if (x > 0 && !diffW && s.getWeight(x) != s.getWeight(x-1)) {
+								diffW = true;
+							}
+							if (x > 0 && !diffS0 && s.getSetupTime(x) != s.getSetupTime(x-1)) {
+								diffS0 = true;
+							}
+							if (x > 0 && !diffD && s.getDueDate(x) != s.getDueDate(x-1)) {
+								diffD = true;
 							}
 						}
[…]
