### 2020-07-15 53093d4
refactored generation & edited docs
