### 2020-07-15 5fca224
Added SingleMachineSchedulingProblem interface
