### 2020-07-15 cicirello/Chips-n-Salsa#9 @nnadeau
Examples should be relevant and applicable
While the examples do a good job of answering the questions of "what" and "how", they do not answer "why".

It would be great if the examples were based on some real-world problems that demonstrate application context and why a dev should use this tool.

Good examples include:
- [`protobuf`](https://developers.google.com/protocol-buffers/docs/javatutorial)
- [`tensorflow`](https://www.tensorflow.org/overview)
- [`pybotics`](https://github.com/nnadeau/pybotics/blob/master/examples/calibration.ipynb)

cc https://github.com/openjournals/joss-reviews/issues/2448
