### 2020-07-15 a91268a
fixed #8
Added a utility class shared by all of the examples.
