### 2020-07-15 9fd07f5
Updated README to fix #3
Added minimum Java requirements.  fixes #3
