### 2020-07-15 cicirello/Chips-n-Salsa#3 @cicirello
@nnadeau  Is something like this what you had in mind? commit 9fd07f538ba52d45a302bbc21b60dae1bc3b8c80
