### 2020-07-15 016f713
update examples README
Updated the examples README as part of #2
