### 2020-07-15 52aa38c
Update README for #2
Updated the readme for issue #2
