+++ README.md
@@ -27,34 +27,41 @@ The GitHub repository is organized as follows:
 
 The library supports Java 8 or higher, including both Oracle JDK 8 and OpenJDK 8.
 
+## Building the Library and Running Examples (Quickstart)
+
+The root of the repository contains a makefile that can be used to: (a) execute the build process, and (b) run 
+all of the example programs back-to-back. The build process uses an ant build file, so the build will fail if 
+you don't have ant installed on your system.  Assuming that you have ant, you can build the library and examples 
+by executing either `make` or `make build` from the root of the repository.  Once the build is complete, you 
+can run all of the examples back-to-back with: `make examples`.  To run a specific example program, see the contents 
+of the makefile, or see the README in the /examples folder.  The following sections contain more detail on
+building the library and running the examples.
+
 ## Building the Library (with ant)
 
-The /build directory contains an ant build file.  The build file also executes the JUnit tests, and the build directory contains the relevant jar files for the JUnit libraries. If you prefer, you can replace these with the latest versions available from https://junit.org/junit4/ (just be sure to edit the property fields in the build.xml to point to the locations of the JUnit jar files).
+The /build directory contains an ant build file.  The build file also executes the JUnit tests, and the build 
+directory contains the relevant jar files for the JUnit libraries. If you prefer, you can replace these 
+with the latest versions available from https://junit.org/junit4/ (just be sure to edit the property 
+fields in the build.xml to point to the locations of the JUnit jar files).
 
-To execute the build process do one of the following. If your working directory is the build directory, then simply execute `ant` from the command line. If your working directory is the parent of build, then execute: `ant -f build/build.xml`
+To execute the build process do one of the following. If your working directory is the build directory, 
+then simply execute `ant` from the command line. If your working directory is the parent of build, 
+then execute: `ant -f build/build.xml`
 
-The default of the provided ant build file, compiles all source files and all JUnit tests, executes all test cases, generates two jar files of the library (one that includes all dependencies, and one that does not), and compiles all example programs. The build process will terminate on any test case failures.  
+The default of the provided ant build file, compiles all source files and all JUnit tests, executes all 
+test cases, generates two jar files of the library (one that includes all dependencies, and one that 
+does not), and compiles all example programs. The build process will terminate on any test case failures.  
 
-The build generates the following directories: bin (for the compiled Java classes), testbin (for the compiled JUnit tests), and exbin (for the compiled example programs).
+The build generates the following directories: bin (for the compiled Java classes), testbin (for the 
+compiled JUnit tests), and exbin (for the compiled example programs).
 
 ## Running the Example Programs
 
-The examples directory contains example usage of several of the classes of the library. The examples directory contains a README file that provides more details on the examples. Each of the examples contains detailed comments within the source code explaining the example. Running the examples without reading the source comments is not advised.
-
-Assuming the build was successful, the compiled example programs will be in the exbin directory.  If your working directory is the root of the repository, then you can run these programs from the command line with the following:
-
-```
-java -cp "exbin;dist/chips-n-salsa-1.0-with-dependencies" org.cicirello.examples.chipsnsalsa.BitVectorExample
-java -cp "exbin;dist/chips-n-salsa-1.0-with-dependencies" org.cicirello.examples.chipsnsalsa.IntegerVectorExample
[…]
