### 2020-07-15 cicirello/Chips-n-Salsa#2 @cicirello
@nnadeau I merged your makefile, and then edited the two READMEs (see the referenced commits) to remove the redundancy.  How does it look now?
