### 2020-07-15 e611d78
Added CI badge
Added CI badge (Github Actions) to README.  #4
