### 2020-07-15 cicirello/Chips-n-Salsa#4 @cicirello
Set up testing/CI with GitHub Actions, in commit 9293c6a9076828314ea046641ba8df013c575b70
