### 2020-07-15 cicirello/Chips-n-Salsa#9 @cicirello
@nnadeau Would it be useful if I included a couple examples that use the library for scheduling problems?
