### 2020-07-15 a7f813e
Minor edit to Makefile
Added a clean to Makefile
