### 2020-07-16 cicirello/Chips-n-Salsa#9 @nnadeau
@cicirello , yes definitely! Some examples from your work/research as well, maybe?
