### 2020-07-16 cicirello/Chips-n-Salsa#2 @nnadeau
@cicirello
- with a `Makefile`, a user should just need to call `make examples` to run all examples. This avoids:

>  Just replace org.cicirello.examples.chipsnsalsa.BitVectorExample in the above with the example program of your [..]

- Modern software development calls for protecting your master branch and using Pull Requests (PR) for all changes. The issue will then be closed by the PR. You can protect master by going to `https://github.com/nnadeau/Chips-n-Salsa/settings/branches` and add master as a protected branch

![image](https://user-images.githubusercontent.com/6395915/87698079-fa660800-c760-11ea-8438-e0ba16b34efb.png)
