### 2020-07-16 acee619
updated examples README
