### 2020-07-16 ddddf66
Added several scheduling problem implementations
