### 2020-07-17 0597d97
Industrial scheduling example #9
Added a detailed example that uses the library to optimize an instance of an industrial scheduling problem.  The example is based on the experiments of my paper from SoCS 2017, and involved the run length schedules from that paper which vary the run length of simulated annealing across parallel restarts.
