+++ examples/org/cicirello/examples/chipsnsalsa/SchedulingExample.java
@@ -1,0 +1,284 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+ 
+package org.cicirello.examples.chipsnsalsa;
+
+import static org.cicirello.examples.chipsnsalsa.ExamplesShared.*;
+import org.cicirello.search.problems.scheduling.SingleMachineSchedulingProblemData;
+import org.cicirello.search.problems.scheduling.WeightedStaticSchedulingWithSetups;
+import org.cicirello.search.problems.scheduling.SingleMachineSchedulingProblem;
+import org.cicirello.search.problems.scheduling.WeightedTardiness;
+import org.cicirello.search.concurrent.TimedParallelMultistarter;
+import org.cicirello.permutations.Permutation;
+import org.cicirello.search.operators.permutations.PermutationInitializer;
+import org.cicirello.search.operators.permutations.InsertionMutation;
+import org.cicirello.search.restarts.ParallelVariableAnnealingLength;
+import org.cicirello.search.restarts.VariableAnnealingLength;
+import org.cicirello.search.sa.SimulatedAnnealing;
+import org.cicirello.search.SolutionCostPair;
+import java.util.ArrayList;
+
+/**
+ * This example program optimizes an instance of an
+ * industrial scheduling problem with jobs that have
+ * sequence-dependent setup times (e.g., setup time required
+ * prior to processing a job depends upon the job that
+ * immediately precedes it on the machine), weights (which
+ * indicate the importance of the job), and due dates.
+ * The cost function minimized in the example is weighted tardiness.
+ * This problem is NP-Hard even without the setup times, which
+ * significantly increase problem solving difficulty.
+ *
+ * This example is based on experiments from a research paper:
+ *       Vincent A. Cicirello. Variable Annealing Length and 
+ *       Parallelism in Simulated Annealing.  Proceedings of the 
+ *       Tenth International Symposium on Combinatorial Search 
+ *       (SoCS 2017), pages 2-10. AAAI Press, June 2017.
+ * That paper proposes two restart schedules for simulated annealing,
+ * Variable Annealing Length (VAL) which varies the length of the runs
+ * for the restarts rather than all restarts having a fixed length.
+ * The run lengths increase exponentially, and the rationale is based on
+ * a commonly encountered phenomena, that a single long run usually outperforms
+ * multiple shorter restarts. However, if we end a longer run early (e.g., if
+ * we run out of time), the result is usually inferior to if the annealing
+ * schedule (i.e., how the temperate varies) was configured for that shorter
+ * run length.  VAL tries to balance the run length tradeoff by starting with
+ * a short run (tuned for that short length), and progressively increasing the 
+ * run length.  The second restart schedule of that paper, P-VAL, spreads the
+ * run length schedule of VAL across multiple parallel multistart searches.
+ *
+ * See the comments within the code for detailed explanations of how the
[…]
