+++ examples/README.md
@@ -32,6 +32,10 @@ integrated (e.g., combining hill climbing with simulated annealing):
 * org.cicirello.examples.chipsnsalsa.PostHillclimbExample
 * org.cicirello.examples.chipsnsalsa.PreHillclimbExample
 
+__Industrial Scheduling Problem Examples__: The following examples use scheduling 
+problems, and are also a bit more complex than some of the other example programs.
+* org.cicirello.examples.chipsnsalsa.SchedulingExample
+
 ## Building the Examples
 
 See the README in the root of the repository (in the parent of this directory) for more details. 
