### 2020-07-21 cicirello/Chips-n-Salsa#7 @cicirello
Now using Github releases (attached latest jars to initial release.
