### 2020-07-21 9878e01
Citation updates #5
1) Replaced some older citations with newer ones.

2) Added a couple newer citations to compliment older ones (where I kept the older ones as first reference of technique).

3) Went through self-cites and kept only those whose content is explicitly mentioned in text (which is most of them as the library is a result of my research).

Fixes #5
