+++ paper/paper.bib
@@ -1,13 +1,3 @@
-@INPROCEEDINGS{Cicirello2019,
-author = {Vincent A. Cicirello},
-title = {Classification of Permutation Distance Metrics for Fitness Landscape Analysis},
-booktitle = {Proceedings of the 11th International Conference on Bio-inspired Information and Communication Technologies},
-publisher = {Springer Nature},
-month = {March},
-pages = {81--97},
-year = {2019},
-doi = {10.1007/978-3-030-24202-2_7}
-}
 
 @article{Cicirello2018b,
 author = {Vincent A. Cicirello},
@@ -54,15 +44,6 @@ pages = {434--446},
 doi = {10.1109/TEVC.2015.2477284}
 }
 
-@INPROCEEDINGS{Cicirello2015,
-author = {Vincent A. Cicirello},
-title = {Genetic Algorithm Parameter Control: Application to Scheduling with Sequence-Dependent Setups},
-booktitle = {Proceedings of the 9th International Conference on Bio-inspired Information and Communications Technologies},
-month = {December},
-year = {2015},
-pages = {136--143},
-doi = {10.4108/eai.3-12-2015.2262346}
-}
 
 @INPROCEEDINGS{Cicirello2014,
 author = {Vincent A. Cicirello},
@@ -85,17 +66,6 @@ year = {2013},
 pages = {46--51}
 }
 
-@INPROCEEDINGS{Cicirello2010,
-author = {Vincent A. Cicirello},
-title = {Heuristic Sequencing Crossover: Integrating Problem Dependent 
-         Heuristic Knowledge into a Genetic Algorithm},
-booktitle = {Proceedings of the Twenty-Third International 
-             Florida Artificial Intelligence Research Society Conference, FLAIRS-23},
-publisher = {AAAI Press},
-month = {May},
-year = {2010},
-pages = {14--19}
-}
 
 @INPROCEEDINGS{Cicirello2007,
 author = {Vincent A. Cicirello},
@@ -107,18 +77,7 @@ month = {September},
 publisher = {AAAI Press}
 }
 
-@INPROCEEDINGS{Cicirello2006,
-author = {Vincent A. Cicirello},
-title = {Non-Wrapping Order Crossover: An Order Preserving
-Crossover Operator that Respects Absolute Position},
-booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference (GECCO'06)},
-year = {2006},
-month = {July},
-pages = {1125--1131},
-volume = {2},
-publisher = {ACM Press},
-doi = {10.1145/1143997.1144177}
-}
+
 
 @ARTICLE{Cicirello2005,
 AUTHOR = {Vincent A. Cicirello and Stephen F. Smith},
@@ -150,36 +109,42 @@ MONTH = {July}
   isbn = {0716710455}
 }
 
-@book{Hoos2005,
-  author    = {Holger H. Hoos and
+@incollection{Hoos2018,
+  title={Stochastic Local Search},
+  author={Holger H. Hoos and
                Thomas St{\"{u}}tzle},
-  title     = {Stochastic Local Search: Foundations and Applications},
-  publisher = {Morgan Kaufmann},
-  year      = {2005},
-  isbn = {1558608729}
+  chapter = {17},
+  year={2018},
+  booktitle={Handbook of Approximation Algorithms and 
+  Metaheuristics Methologies and Traditional Applications}, 
+  volume={1},
+  edition={2nd},
+  publisher ={Chapman and Hall/CRC}
 }
 
-@article {Kirkpatrick1983,
-	author = {Kirkpatrick, S. and Gelatt, C. D. and Vecchi, M. P.},
-	title = {Optimization by Simulated Annealing},
-	volume = {220},
-	number = {4598},
-	pages = {671--680},
[…]
