+++ paper/paper.bib
@@ -193,7 +193,8 @@ pages = {850--858}
   isbn={9781848218048},
   series={Computer Engineering: Metaheuristics},
   year={2017},
-  publisher={Wiley}
+  publisher={Wiley},
+  doi={https://doi.org/10.1002/9781119136378}
 }
 
 @book{Doerr2019,
