### 2020-07-21 #2448 @cicirello
@galessiorob I just made a few edits to the paper.md and paper.tex.  When generating the pdf with whedon, I got a very strange error that appears to be related to one of the reviewer ids (see the part of the error that says: "pandoc-citeproc: reference mbdemoraes not found"), and not at all due to the paper.  See the error in the result from whedon above.  I just tried the preview service (https://whedon.theoj.org/) to confirm that it isn't due to paper.md or paper.tex.  The paper compiles fine with the preview service.

Any idea how to correct this?
