### 2020-07-22 664612a
Badge link to releases
Added link to releases from release badge
