### 2020-07-22 618e56e
workflow publish to GPR
Added workflow to publish to GitHub Package Registry whenever new release.
