### 2020-07-23 5d5abdd
refactored constructor
