+++ src/org/cicirello/search/hc/SteepestDescentHillClimber.java
@@ -230,8 +230,7 @@ public final class SteepestDescentHillClimber<T extends Copyable<T>> implements 
 	}
 	
 	private OneClimb<T> initClimberInt() {
-		return new OneClimb<T>() {
-			public SolutionCostPair<T> climbOnce(T current) {
+		return current -> {
 				// compute cost of start
 				int currentCost = pOptInt.cost(current);
 				boolean keepClimbing = true;
@@ -262,13 +261,11 @@ public final class SteepestDescentHillClimber<T extends Copyable<T>> implements 
 					}
 				}
 				return new SolutionCostPair<T>(current, currentCost);
-			}
 		};
 	}
 	
 	private OneClimb<T> initClimberDouble() {
-		return new OneClimb<T>() {
-			public SolutionCostPair<T> climbOnce(T current) {
+		return current -> {
 				// compute cost of start
 				double currentCost = pOpt.cost(current);				
 				boolean keepClimbing = true;
@@ -299,7 +296,6 @@ public final class SteepestDescentHillClimber<T extends Copyable<T>> implements 
 					}
 				}
 				return new SolutionCostPair<T>(current, currentCost);
-			}
 		};
 	}
 	
