### 2020-07-23 c91b6ff
refactor JOSS paper text
