### 2020-07-23 cdd9573
refactored JOS paper text
