+++ paper/paper.md
@@ -21,20 +21,20 @@ bibliography: paper.bib
 
 # Summary
 
-Discrete optimization problems are often both of practical importance in the real-world as well as computationally intractable. For example, the optimization variants of the traveling salesperson, bin packing, and longest common subsequence problems are NP-Hard, as are the resource constrained scheduling problem, and even many single-machine sequencing problems [@Garey1979]. Polynomial time algorithms for such problems are unlikely to exist, and the best known algorithms that are guaranteed to optimally solve these problems have a worst-case exponential runtime. It is thus common to turn to stochastic local search and other forms of metaheuristic [@Gonzalez2018] when dealing with these and related problems. Stochastic local search algorithms begin at a randomly generated search state, and then apply a sequence of neighbor transitions to nearby similar search states. Stochastic local search includes perturbative [@Hoos2018] algorithms like simulated annealing [@Delahaye2019] and hill climbers [@Hoos2018], where each search state is a complete candidate feasible solution, and a mutation operator (or neighborhood operator) makes a small modification (usually randomly) to move to another local candidate solution; and also includes constructive [@Hoos2018] algorithms like stochastic samplers [@Grasas2017; @ReyesRubiano2020; @Cicirello2005; @Bresina1996; @Langley1992], where each search state is a partial solution that is iteratively transformed into a complete feasible solution. Stochastic local search algorithms are not guaranteed to find optimal solutions. However, they often find sufficiently optimal solutions in much less time than systematic search. They also are characterized by an anytime property [@Jesus2020; @Zilberstein1996], such that the quality of the solution improves with runtime.
+Discrete optimization problems are often both of practical real-world importance as well as computationally intractable. For example, the optimization variants of the traveling salesperson, bin packing, and longest common subsequence problems are NP-Hard, as are the resource constrained scheduling problem, and even many single-machine sequencing problems [@Garey1979]. Polynomial time algorithms for such problems are unlikely to exist, and the best known algorithms that are guaranteed to optimally solve these problems have a worst-case exponential runtime. It is thus common to turn to stochastic local search and other forms of metaheuristic [@Gonzalez2018] when dealing with these and related problems. Stochastic local search algorithms begin at a random search state, and then apply a sequence of neighbor transitions to nearby similar search states. Stochastic local search includes perturbative [@Hoos2018] algorithms like simulated annealing [@Delahaye2019] and hill climbers [@Hoos2018], where each search state is a complete candidate feasible solution, and a mutation operator (or neighborhood operator) makes a small modification (usually randomly) to move to another local candidate solution; and also includes constructive [@Hoos2018] algorithms like stochastic samplers [@Grasas2017; @ReyesRubiano2020; @Cicirello2005; @Bresina1996; @Langley1992], where each search state is a partial solution that is iteratively transformed into a complete feasible solution. Stochastic local search algorithms are not guaranteed to find optimal solutions. However, they often find sufficiently optimal solutions in much less time than systematic search. They also are characterized by an anytime property [@Jesus2020; @Zilberstein1996], such that the quality of the solution improves with runtime.
 
[…]
