### 2020-07-23 1814bc6
JOSS paper refactoring
