### 2020-07-23 5f103d1
JOSS paper refactorings
