### 2020-07-23 595eb6d
More JOSS text refactorings
