+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 Discrete optimization problems are often both of practical real-world importance as well as computationally intractable. For example, the optimization variants of the traveling salesperson, bin packing, and longest common subsequence problems are NP-Hard, as is resource constrained scheduling, and even many single-machine scheduling problems [@Garey1979]. Polynomial time algorithms for such problems are unlikely to exist, and the best known algorithms that are guaranteed to optimally solve these problems have a worst-case exponential runtime. It is thus common to use stochastic local search and other metaheuristics [@Gonzalez2018] for these and related problems. Stochastic local search algorithms begin at a random search state, and apply a sequence of neighbor transitions to nearby similar search states. This includes perturbative [@Hoos2018] algorithms like simulated annealing [@Delahaye2019] and hill climbers [@Hoos2018], where each search state is a complete candidate feasible solution, and a mutation operator (or neighborhood operator) makes a small modification (usually randomly) to move to another local candidate solution; and also includes constructive [@Hoos2018] algorithms like stochastic samplers [@Grasas2017; @ReyesRubiano2020; @Cicirello2005; @Bresina1996; @Langley1992], where each search state is a partial solution that is iteratively transformed into a complete feasible solution. Stochastic local search algorithms are not guaranteed to find optimal solutions. However, they often find near-optimal solutions in much less time than systematic search. They also are characterized by an anytime property [@Jesus2020; @Zilberstein1996], such that the quality of the solution improves with runtime.
 
-Chips-n-Salsa is a Java library of customizable, hybridizable, iterative, parallel, stochastic, and self-adaptive local search algorithms. Its focus is discrete optimization, but also supports continuous optimization. The library provides a variety of solution representations. For example, it includes BitVector and IntegerVector classes, indexable vectors of bits and integers, respectively, along with mutation operators for use by local search for these types. The library also utilizes our significant prior research on permutation optimization problems, and provides an extensive set of mutation operators for permutations [@Cicirello2016; @Cicirello2013], including window-limited mutation [@Cicirello2014] operators. It uses the efficiently implemented Permutation class of the JavaPermutationTools (JPT) [@Cicirello2018b] library for representing solutions to such problems. For continuous problems, Chips-n-Salsa provides a RealVector class, Gaussian mutation [@Petrowski2017], Cauchy mutation [@Petrowski2017], and uniform mutation. The library also includes optimization problems useful for benchmarking new metaheuristic implementations, such as the well-known OneMax problem [@Doerr2019], BoundMax (generalization of OneMax to integers), the Permutation in a Haystack problem [@Cicirello2016], and polynomial root finding.
[…]
