+++ paper/paper.md
@@ -25,7 +25,7 @@ Discrete optimization problems are often both of practical real-world importance
 
 Chips-n-Salsa is a Java library of customizable, hybridizable, iterative, parallel, stochastic, and self-adaptive local search algorithms. Its focus is discrete optimization, but also supports continuous optimization. The library provides a variety of solution representations. For example, it includes BitVector and IntegerVector classes, indexable vectors of bits and integers, respectively, along with mutation operators for these types. The library also utilizes our significant prior research on permutation optimization problems, and provides an extensive set of mutation operators for permutations [@Cicirello2016; @Cicirello2013], including window-limited mutation [@Cicirello2014]. It uses the efficiently implemented Permutation class of the JavaPermutationTools (JPT) [@Cicirello2018b] library for representing solutions to such problems. For continuous problems, Chips-n-Salsa provides a RealVector class, Gaussian mutation [@Petrowski2017], Cauchy mutation [@Petrowski2017], and uniform mutation. The library also includes optimization problems useful for benchmarking new metaheuristic implementations, such as the well-known OneMax problem [@Doerr2019], BoundMax (generalization of OneMax to integers), the Permutation in a Haystack problem [@Cicirello2016], and polynomial root finding.
 
-The repository (https://github.com/cicirello/Chips-n-Salsa) contains source code of the library, and programs that provide example usage of key functionality. API and other documentation is hosted on the web (https://chips-n-salsa.cicirello.org/). 
+The repository (https://github.com/cicirello/Chips-n-Salsa) contains source code of the library, and programs that provide example usage of key functionality. API and other documentation is hosted on the web (https://chips-n-salsa.cicirello.org/). The library can be integrated into projects using maven via GitHub Packages (https://github.com/cicirello/Chips-n-Salsa/packages).
 
 # Chips-n-Salsa Features and Functionality:
 
