### 2020-07-23 2d37093
Rewrote statement of need
Rewrote statement of need section to discuss other related libraries.  Fixes #6
