+++ paper/paper.bib
@@ -313,7 +313,34 @@ booktitle = {Proceedings of the Fifteenth National Conference on Artificial Inte
 pages = {431–-437}
 }
 
+@misc{ECJ,
+  author = {​Sean Luke},
+  title = {{ECJ} Evolutionary Computation Toolkit},
+  year = {2020},
+  publisher = {​GitHub},
+  journal = {​GitHub repository},
+  url = {​https://github.com/GMUEClab/ecj}
+}
 
+@misc{Jenetics,
+  author = {​Jenetics},
+  title = {Jenetics - Genetic Algorithm, Genetic Programming, Evolutionary Algorithm, and Multi-objective Optimization},
+  year = {2020},
+  publisher = {​GitHub},
+  journal = {​GitHub repository},
+  url = {​https://github.com/jenetics/jenetics}
+}
+
+@article{Pagnozzi2019,
+title = {Automatic design of hybrid stochastic local search algorithms for permutation flowshop problems},
+journal = {European Journal of Operational Research},
+volume = {276},
+number = {2},
+pages = {409--421},
+year = {2019},
+doi = {10.1016/j.ejor.2019.01.018},
+author = {Federico Pagnozzi and Thomas Stützle}
+}
 
 
 
