+++ paper/paper.md
@@ -38,6 +38,8 @@ The repository (https://github.com/cicirello/Chips-n-Salsa) contains source code
 
 # Statement of Need
 
-The target audience of Chips-n-Salsa includes those conducting computational research in diverse domains, wherever applications of NP-Hard optimization problems is important. Existing local search and metaheuristic implementations that are freely available are often intimately tied to a specific problem, or at best to a specific representation; whereas Chips-n-Salsa is not only problem-independent, but also representation-independent. Its support for parallel execution enables easily exploiting multicore processor architectures to speed up problem solving.  The self-adaptive and self-tuning features streamline the development process, eliminating the often tedious stage, control parameter optimization, of applying stochastic local search to a new problem.
+The target audience of Chips-n-Salsa includes those conducting computational research in diverse domains, wherever applications of NP-Hard optimization problems is important. Existing local search and metaheuristic open source implementations are often intimately tied to a specific problem, and usually a specific form of local search. For example, a GitHub search finds countless solvers for problems such as Boolean Satisfiability or the Traveling Salesperson. Such problem-dependent implementations are not easily adapted to new problems. There are some notable exceptions. In particular, ECJ [@ECJ] and Jenetics [@Jenetics] are both mature and well-maintained Java libraries for genetic algorithms and related forms of evolutionary computation. They are problem-independent, support multiple representations, and include multithreaded capabilities. Their focus, however, is on population-based evolutionary search. Another especially noteworthy project is emili [@Pagnozzi2019], which is a C++ framework supporting hybrid stochastic local search algorithms. 
+
+Chips-n-Salsa is not only problem-independent, but also representation-independent. Its support for parallel execution enables easily exploiting multicore processor architectures to speed up problem solving.  The self-adaptive and self-tuning features streamline the development process, eliminating an often tedious stage, control parameter optimization, of applying stochastic local search to a new problem.
 
 # References
