### 2020-07-23 faf7c61
refactored JOSS paper text
After adding discussion of related libraries, paper was over 1000 word limit. Edited and refactored text to bring down below the limit.
