### 2020-07-23 697e6ae
Shortened paper a bit more
