+++ paper/paper.bib
@@ -1,4 +1,3 @@
-
 @article{Cicirello2018b,
 author = {Vincent A. Cicirello},
 title = {{JavaPermutationTools}: A Java Library of Permutation Distance Metrics},
@@ -318,8 +317,7 @@ pages = {431–-437}
   title = {{ECJ} Evolutionary Computation Toolkit},
   year = {2020},
   publisher = {​GitHub},
-  journal = {​GitHub repository},
-  url = {​https://github.com/GMUEClab/ecj}
+  journal = {​GitHub repository}
 }
 
 @misc{Jenetics,
