+++ paper/paper.bib
@@ -325,8 +325,7 @@ pages = {431–-437}
   title = {Jenetics - Genetic Algorithm, Genetic Programming, Evolutionary Algorithm, and Multi-objective Optimization},
   year = {2020},
   publisher = {​GitHub},
-  journal = {​GitHub repository},
-  url = {​https://github.com/jenetics/jenetics}
+  journal = {​GitHub repository}
 }
 
 @article{Pagnozzi2019,
