+++ paper/paper.md
@@ -38,8 +38,6 @@ The repository (https://github.com/cicirello/Chips-n-Salsa) contains the library
 
 # Statement of Need
 
-The target audience of Chips-n-Salsa includes those conducting computational research in diverse domains, wherever applications of NP-Hard optimization problems is important. Existing local search open source implementations are often intimately tied to a specific problem, and usually a specific form of local search. For example, a GitHub search finds countless solvers for problems like Boolean Satisfiability or the Traveling Salesperson. Such problem-dependent implementations are not easily adapted to new problems. There are notable exceptions. In particular, ECJ [@ECJ] and Jenetics [@Jenetics] are both mature and well-maintained Java libraries for genetic algorithms and related forms of evolutionary computation. They are problem-independent, support multiple representations, and include multithreaded capabilities. Their focus, however, is on population-based evolutionary search. Another especially noteworthy project is emili [@Pagnozzi2019], which is a C++ framework supporting hybrid stochastic local search algorithms. 
-
-Chips-n-Salsa is not only problem-independent, but also representation-independent, and supports hybrid local search, as well as parallel execution to easily exploit multicore architectures to speed up problem solving. The self-adaptive and self-tuning features streamline development, eliminating the need for the developer to tune control parameters.
+The target audience of Chips-n-Salsa includes those conducting computational research in diverse domains, wherever applications of NP-Hard optimization problems is important. Existing local search open source implementations are often intimately tied to a specific problem, and usually a specific form of local search. For example, a GitHub search finds countless solvers for problems like Boolean Satisfiability or the Traveling Salesperson. Such problem-dependent implementations are not easily adapted to new problems. There are notable exceptions. In particular, ECJ [@ECJ] and Jenetics [@Jenetics] are both mature and well-maintained Java libraries for genetic algorithms and related forms of evolutionary computation. They are problem-independent, support multiple representations, and include multithreaded capabilities. Their focus, however, is on population-based evolutionary search; whereas Chips-n-Salsa is focused on single-solution algorithms. Another especially noteworthy project is emili [@Pagnozzi2019], which is a C++ framework supporting hybrid, single-solution, stochastic local search algorithms. Chips-n-Salsa is problem-independent, representation-independent, and supports hybrid local search, as well as parallel execution to easily exploit multicore architectures to speed up problem solving. Furthermore, the self-adaptive and self-tuning features streamline development, eliminating the need for the developer to tune control parameters.
 
 # References
