+++ paper/paper.bib
@@ -312,12 +312,14 @@ booktitle = {Proceedings of the Fifteenth National Conference on Artificial Inte
 pages = {431–-437}
 }
 
-@misc{ECJ,
-  author = {​Sean Luke},
-  title = {{ECJ} Evolutionary Computation Toolkit},
-  year = {2020},
-  publisher = {​GitHub},
-  journal = {​GitHub repository}
+@inproceedings{Scott2019,
+author = {Scott, Eric O. and Luke, Sean},
+title = {ECJ at 20: Toward a General Metaheuristics Toolkit},
+year = {2019},
+publisher = {ACM Press},
+doi = {10.1145/3319619.3326865},
+booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference Companion},
+pages = {1391--1398}
 }
 
 @misc{Jenetics,
@@ -325,7 +327,8 @@ pages = {431–-437}
   title = {Jenetics - Genetic Algorithm, Genetic Programming, Evolutionary Algorithm, and Multi-objective Optimization},
   year = {2020},
   publisher = {​GitHub},
-  journal = {​GitHub repository}
+  journal = {​GitHub repository},
+  url = {https://jenetics.io/}
 }
 
 @article{Pagnozzi2019,
