### 2020-07-23 cicirello/Chips-n-Salsa#6 @cicirello
@nnadeau  I think I've addressed all of your feedback so far.  To see the pdf of the current version of the paper, use the preview service (https://whedon.theoj.org/). The whedon pdf generator is giving a weird error that they are investigating if generated from in the review issue, but it generates fine from the preview service.  Error appears related to something in the review issue itself, rather than the paper.
