### 2020-07-24 522adf4
Update ConstructiveHeuristic.java
