### 2020-07-24 c83fb82
Create SchedulingHeuristic.java
Serves as an abstract base class for the scheduling heuristics.
