### 2020-07-24 e5a7b3d
Testcase for EDD heuristic
