+++ tests/org/cicirello/search/operators/integers/UniformMutationTests.java
@@ -40,7 +40,7 @@ public class UniformMutationTests {
 	// increases value, and if at least 1 out of MAX_TRIALS decreases.
 	// A Uniform is symmetric about 0.0, so approximately half of mutations 
 	// should decrease and approximately half should increase.
-	private static final int MAX_TRIALS = 40;
+	private static final int MAX_TRIALS = 100;
 	
 	@Test
 	public void testUniformMutation1() {
