### 2020-07-24 8906951
tests for WSPT heuristic
