+++ tests/org/cicirello/search/problems/scheduling/HeuristicTests.java
@@ -43,6 +43,20 @@ public class HeuristicTests {
 		}
 	}
 	
+	@Test
+	public void testWSPT() {
+		double e = WeightedShortestProcessingTime.EPSILON;
+		int highP = (int)Math.ceil(1 / e)*2;
+		int[] w =    { 1, 1, 1, 1, 0, 0, 0, 0, 2, 2, 2, 2, 1};
+		int[] p =    { 1, 2, 4, 8, 1, 2, 4, 8, 1, 2, 4, 8, highP};
+		double[] expected = { 1, 0.5, 0.25, 0.125, e, e, e, e, 2, 1, 0.5, 0.25, e};
+		FakeProblemWeightsPTime problem = new FakeProblemWeightsPTime(w, p);
+		WeightedShortestProcessingTime h = new WeightedShortestProcessingTime(problem);
+		for (int j = 0; j < expected.length; j++) {
+			assertEquals(expected[j], h.h(null, j, null), 1E-10);
+		}
+	}
+	
 	private static class FakeProblemDuedates implements SingleMachineSchedulingProblem {
 		
 		private FakeProblemData data;
@@ -60,18 +74,44 @@ public class HeuristicTests {
 		@Override public int value(Permutation p) { return 10; }
 	}
 	
+	private static class FakeProblemWeightsPTime implements SingleMachineSchedulingProblem {
+		
+		private FakeProblemData data;
+		
+		public FakeProblemWeightsPTime(int[] w, int[] p) {
+			data = new FakeProblemData(w, p);
+		}
+		
+		@Override
+		public SingleMachineSchedulingProblemData getInstanceData() {
+			return data;
+		}
+		
+		@Override public int cost(Permutation p) { return 10; }
+		@Override public int value(Permutation p) { return 10; }
+	}
+	
 	private static class FakeProblemData implements SingleMachineSchedulingProblemData {
 
 		private int[] d;
+		private int[] w;
+		private int[] p;
 		
 		public FakeProblemData(int[] d) {
-			this.d = d;
+			this.d = d.clone();
+		}
+		
+		public FakeProblemData(int[] w, int[] p) {
+			this.w = w.clone();
+			this.p = p.clone();
 		}
 		
 		@Override public int numberOfJobs() { return d.length; }
-		@Override public int getProcessingTime(int j) { return 0; }
+		@Override public int getProcessingTime(int j) { return p==null? 0: p[j]; }
 		@Override public int[] getCompletionTimes(Permutation schedule) { return null; }
-		@Override public boolean hasDueDates() { return true; }
+		@Override public boolean hasDueDates() { return d != null; }
 		@Override public int getDueDate(int j) { return d[j]; }
+		@Override public boolean hasWeights() { return w != null; }
+		@Override public int getWeight(int j) { return w[j]; }
 	}	
 }
