### 2020-07-24 d7f94a0
Added variation of WSPT heuristic
