### 2020-07-24 b858266
tests and fix for WSPT2
