+++ tests/org/cicirello/search/problems/scheduling/HeuristicTests.java
@@ -57,6 +57,67 @@ public class HeuristicTests {
 		}
 	}
 	
+	@Test
+	public void testWSPT2() {
+		double e = WeightedShortestProcessingTimeLateOnly.EPSILON;
+		int highP = (int)Math.ceil(1 / e)*2;
+		int[] w =    { 1, 1, 1, 1, 1, 0, 0, 0, 0, 2, 2, 2, 2, 1};
+		int[] p =    { 1, 1, 2, 4, 8, 1, 2, 4, 8, 1, 2, 4, 8, highP};
+		double[] expected = { 999, 1, 0.5, 0.25, 0.125, e, e, e, e, 2, 1, 0.5, 0.25, e};
+		PartialPermutation partial = new PartialPermutation(expected.length);
+		//Doesn't really matter: partial.extend(0);
+		// All late tests
+		FakeProblemWeightsPTime problem = new FakeProblemWeightsPTime(w, p, 0);
+		WeightedShortestProcessingTimeLateOnly h = new WeightedShortestProcessingTimeLateOnly(problem);
+		IncrementalEvaluation inc = h.createIncrementalEvaluation();
+		inc.extend(partial, 0);
+		for (int j = 1; j < expected.length; j++) {
+			assertEquals("j:"+j, expected[j], h.h(partial, j, inc), 1E-10);
+		}
+		// All on time tests
+		problem = new FakeProblemWeightsPTime(w, p, 20);
+		h = new WeightedShortestProcessingTimeLateOnly(problem);
+		inc = h.createIncrementalEvaluation();
+		inc.extend(partial, 0);
+		for (int j = 1; j < expected.length; j++) {
+			assertEquals(e, h.h(partial, j, inc), 1E-10);
+		}
+	}
+	
+	@Test
+	public void testWSPT2IncEvalExtend() {
+		int[] w = { 1, 1, 1, 1, 1 };
+		int[] p = { 3, 2, 1, 4, 5 };
+		int[] e = { 3, 5, 6, 10, 15 };
+		FakeProblemWeightsPTime problem = new FakeProblemWeightsPTime(w, p, 0);
+		WeightedShortestProcessingTimeLateOnly h = new WeightedShortestProcessingTimeLateOnly(problem);
+		PartialPermutation partial = new PartialPermutation(e.length);
+		WeightedShortestProcessingTimeLateOnly.IncrementalTimeCalculator inc = (WeightedShortestProcessingTimeLateOnly.IncrementalTimeCalculator)h.createIncrementalEvaluation();
+		assertEquals(0, inc.currentTime());
+		for (int i = 0; i < p.length; i++) {
+			inc.extend(partial, i);
+			partial.extend(0);
+			assertEquals(e[i], inc.currentTime());
+		}
+	}
+	
+	@Test
+	public void testWSPT2IncEvalExtendWithSetups() {
+		int[] w = { 1, 1, 1, 1, 1 };
+		int[] p = { 3, 2, 1, 4, 5 };
+		int[] e = { 10, 13, 16, 23, 32 };
+		FakeProblemWeightsPTime problem = new FakeProblemWeightsPTime(w, p, 0, 7);
+		WeightedShortestProcessingTimeLateOnly h = new WeightedShortestProcessingTimeLateOnly(problem);
+		PartialPermutation partial = new PartialPermutation(e.length);
+		WeightedShortestProcessingTimeLateOnly.IncrementalTimeCalculator inc = (WeightedShortestProcessingTimeLateOnly.IncrementalTimeCalculator)h.createIncrementalEvaluation();
+		assertEquals(0, inc.currentTime());
+		for (int i = 0; i < p.length; i++) {
+			inc.extend(partial, i);
+			partial.extend(0);
+			assertEquals(e[i], inc.currentTime());
+		}
+	}
+	
[…]
