### 2020-07-27 a9a8b45
Min heuristic value constant
