### 2020-07-27 3970cdf
EDD lower bound at MIN_H
