### 2020-07-27 b055738
EPSILON renamed MIN_H
