+++ src/org/cicirello/search/problems/scheduling/WeightedShortestProcessingTime.java
@@ -29,7 +29,7 @@ import org.cicirello.search.ss.IncrementalEvaluation;
  * heuristic.  This heuristic is usually defined as: h(j) = w[j] / p[j],
  * where w[j] is the weight of job j, and p[j] is its processing time.
  * This implementation alters this definition slightly as:
- * h(j) = max( EPSILON, w[j] / p[j]), where EPSILON (a class constant)
+ * h(j) = max( {@link #MIN_H}, w[j] / p[j]), where {@link #MIN_H}
  * is a small non-zero value.  This is to deal with the possibility of
  * a job with weight w[j] = 0.  For deterministic construction of a 
  * schedule, this adjustment is unnecessary.  However, for stochastic sampling
@@ -40,11 +40,6 @@ import org.cicirello.search.ss.IncrementalEvaluation;
  * @version 7.24.2020
  */
 public final class WeightedShortestProcessingTime extends SchedulingHeuristic {
-	
-	/**
-	 * The minimum heuristic value.
-	 */
-	public static final double EPSILON = 0.00001;
 	
 	/**
 	 * Constructs an WeightedShortestProcessingTime heuristic.
@@ -58,8 +53,8 @@ public final class WeightedShortestProcessingTime extends SchedulingHeuristic {
 	@Override
 	public double h(PartialPermutation p, int element, IncrementalEvaluation incEval) {
 		double w = data.getWeight(element);
-		if (w <= EPSILON) return EPSILON;
+		if (w <= MIN_H) return MIN_H;
 		double value = w / data.getProcessingTime(element);
-		return value < EPSILON ? EPSILON : value;
+		return value <= MIN_H ? MIN_H : value;
 	}
 }
