### 2020-07-27 1f2320c
EPSILON renamed MIN_H
