### 2020-07-27 f87cd8f
refactored SchedulingHeuristic hierarchy
