### 2020-07-27 e57ede6
refactored SchedulingHeuristic hierarchy
