### 2020-07-27 5360d09
Update HeuristicTests.java
