### 2020-07-27 84f03c6
refactor slack handling
