### 2020-07-27 3d80f67
tests for slack computation
