### 2020-07-27 5f014f7
Weighted Critical Ratio heuristic
