+++ src/org/cicirello/search/problems/scheduling/WeightedCriticalRatio.java
@@ -49,7 +49,7 @@ import org.cicirello.search.ss.PartialPermutation;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 7.24.2020
+ * @version 7.27.2020
  */
 public final class WeightedCriticalRatio extends WeightedShortestProcessingTime {
 	
