### 2020-07-27 724e59a
Montagne's heuristic
