+++ src/org/cicirello/search/problems/scheduling/Montagne.java
@@ -1,0 +1,78 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.cicirello.search.problems.scheduling;
+
+import org.cicirello.permutations.Permutation;
+import org.cicirello.search.ss.IncrementalEvaluation;
+import org.cicirello.search.ss.PartialPermutation;
+
+/**
+ * <p>This is an implementation of Montagne's heuristic
+ * heuristic.  This heuristic is defined as:
+ * h(j) = (w[j]/p[j])(1 - (d[j]/&sigma;<sub>i</sub>p[i]),
+ * where w[j] is the weight of job j, p[j] is its processing time,
+ * and d[j] is job j's due date.  The &sigma;<sub>i</sub>p[i] is
+ * the sum of the processing times of the remaining unscheduled jobs.</p>  
+ *
+ * <p>The constant {@link #MIN_H} defines the minimum value
+ * the heuristic will return, preventing h(j)=0 in support of stochastic
+ * sampling algorithms for which h(j)=0 is problematic.  This implementation 
+ * returns max( {@link #MIN_H}, h(j)), where {@link #MIN_H}
+ * is a small non-zero value.</p>  
+ *
+ * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
+ * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
+ * @version 7.27.2020
+ */
+public final class Montagne extends WeightedShortestProcessingTime {
+	
+	/**
+	 * Constructs an Montagne heuristic.
+	 * @param problem The instance of a scheduling problem that is
+	 * the target of the heuristic.
+	 * @throws IllegalArgumentException if problem.hasDueDates() returns false.
+	 */
+	public Montagne(SingleMachineSchedulingProblem problem) {
+		super(problem);
+		if (!data.hasDueDates()) {
+			throw new IllegalArgumentException("This heuristic requires due dates.");
+		}
+	}
+	
+	@Override
+	public double h(PartialPermutation p, int element, IncrementalEvaluation incEval) {
+		double value = super.h(p, element, incEval);
+		if (value > MIN_H) {
+			value *= (1.0 - 
+				((double)data.getDueDate(element)) /			
+				((IncrementalAverageProcessingCalculator)incEval).totalProcessingTime()
+			);
+			return value <= MIN_H ? MIN_H : value;
+		}		
+		return value;
+	}
+	
+	@Override
+	public IncrementalEvaluation createIncrementalEvaluation() {
+		return new IncrementalAverageProcessingCalculator();
+	}
+	
+}
