### 2020-07-27 123a2c5
SPT heuristic
