### 2020-07-27 4dcfac7
ATC heuristic
