### 2020-07-28 285d599
refactored MST heuristic
