### 2020-07-28 7b8bf07
optimized EDD
