### 2020-07-28 9eb32da
extracted a helper
