### 2020-07-28 bc94c75
optimized MST
