### 2020-07-28 a42877b
optimized SPT heuristic
