### 2020-07-28 8d0594d
optimized WST heuristic
