+++ README.md
@@ -20,13 +20,13 @@ Chips-n-Salsa is customizable, making extensive use of Java's generic types, ena
 ## Repository Organization
 
 The GitHub repository is organized as follows:
-* The /build directory contains an ant build file, and other resources related to building the library.
-* The /dist directory contains the compiled jar files of the library, one consisting of just the Chips-n-Salsa library, and a second jar file that contains Chips-n-Salsa as well as all dependencies.  The file name makes it clear which is which.
-* The /docs directory contains the javadoc documentation in a sub-directory /docs/api. The /docs directory is also the location of the website for the project hosted via GitHub pages at https://chips-n-salsa.cicirello.org/.
-* The /examples directory contains several example programs that use the library, and which demonstrate usage of various features.
-* The /lib directory contains jar files of all dependencies.  At the present time, the only dependency is a library [JavaPermutationTools (JPT)](https://jpt.cicirello.org/) which is used for its representation of a Permutation, as well as for some of its random number generation capabilities.
-* The /src directory contains all of the source code for Chips-n-Salsa.
-* The /tests directory contains JUnit test cases for all functionality of the library.
+* The [/build](build) directory contains an ant build file, and other resources related to building the library.
+* The [/dist](dist) directory contains the compiled jar files of the library, one consisting of just the Chips-n-Salsa library, and a second jar file that contains Chips-n-Salsa as well as all dependencies.  The file name makes it clear which is which.
+* The [/docs](docs) directory contains the javadoc documentation in a sub-directory /docs/api. The /docs directory is also the location of the website for the project hosted via GitHub pages at https://chips-n-salsa.cicirello.org/.
+* The [/examples](examples) directory contains several example programs that use the library, and which demonstrate usage of various features.
+* The [/lib](lib) directory contains jar files of all dependencies.  At the present time, the only dependency is a library [JavaPermutationTools (JPT)](https://jpt.cicirello.org/) which is used for its representation of a Permutation, as well as for some of its random number generation capabilities.
+* The [/src](src) directory contains all of the source code for Chips-n-Salsa.
+* The [/tests](tests) directory contains JUnit test cases for all functionality of the library.
 
 ## Java 8
 
@@ -34,38 +34,38 @@ The library supports Java 8 or higher, including both Oracle JDK 8 and OpenJDK 8
 
 ## Building the Library and Running Examples (Quickstart)
 
-The root of the repository contains a makefile that can be used to: (a) execute the build process, and (b) run 
-all of the example programs back-to-back. The build process uses an ant build file, so the build will fail if 
-you don't have ant installed on your system.  Assuming that you have ant, you can build the library and examples 
-by executing either `make` or `make build` from the root of the repository.  Once the build is complete, you 
-can run all of the examples back-to-back with: `make examples`.  To run a specific example program, see the contents 
+The root of the repository contains a makefile that can be used to: (a) execute the build process, and (b) run
+all of the example programs back-to-back. The build process uses an ant build file, so the build will fail if
+you don't have ant installed on your system.  Assuming that you have ant, you can build the library and examples
+by executing either `make` or `make build` from the root of the repository.  Once the build is complete, you
+can run all of the examples back-to-back with: `make examples`.  To run a specific example program, see the contents
 of the makefile, or see the README in the /examples folder.  The following sections contain more detail on
 building the library and running the examples.
 
 ## Building the Library (with ant)
 
-The /build directory contains an ant build file.  The build file also executes the JUnit tests, and the build 
-directory contains the relevant jar files for the JUnit libraries. If you prefer, you can replace these 
-with the latest versions available from https://junit.org/junit4/ (just be sure to edit the property 
+The /build directory contains an ant build file.  The build file also executes the JUnit tests, and the build
+directory contains the relevant jar files for the JUnit libraries. If you prefer, you can replace these
[…]
