### 2020-07-28 cicirello/Chips-n-Salsa#23 @nnadeau
Update README to include GitHub Packages maven info
Now that GH Packages is being used to distribute the package, there should be a README section that describes how to import this package.

cc https://github.com/openjournals/joss-reviews/issues/2448
