### 2020-07-28 cicirello/Chips-n-Salsa#24 @nnadeau
Define versioning scheme
I noticed that the GH Releases/Packages appear to be using SemVer (https://semver.org/). The README should be updated to include a section that explicitly defines the versioning scheme.

cc https://github.com/openjournals/joss-reviews/issues/2448
