### 2020-07-28 cicirello/Chips-n-Salsa#25 @nnadeau
Add CONTRIBUTING.md
https://docs.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors

cc https://github.com/openjournals/joss-reviews/issues/2448
