### 2020-07-28 #2448 @nnadeau
@cicirello almost there! Great job so far! I believe these should be the last issues:

- https://github.com/cicirello/Chips-n-Salsa/issues/25
- https://github.com/cicirello/Chips-n-Salsa/pull/22
- https://github.com/cicirello/Chips-n-Salsa/issues/24
- https://github.com/cicirello/Chips-n-Salsa/issues/23
