### 2020-07-28 2c5bd49
Added instructions for importing maven package #23
Fixes #23
