+++ README.md
@@ -69,6 +69,36 @@ comments within the source code explaining the example. Running the examples wit
 source comments is not advised.  Assuming the build was successful, the compiled example programs
 will be in the exbin directory.  See /examples/README.md for details on running the examples individually.
 
+## Importing the Library Into a Project Using Maven
+
+__Step 1__: Add this to the dependencies section of your pom.xml, replacing the version number with the version you want to use.
+
+```XML
+<dependency>
+  <groupId>org.cicirello</groupId>
+  <artifactId>chips-n-salsa</artifactId>
+  <version>1.1.0</version>
+</dependency>
+```
+
+__Step 2__: Add this to the repositories section of your pom.xml.
+
+```XML
+<repository>
+  <id>github</id>
+  <name>GitHub cicirello Apache Maven Packages</name>
+  <url>https://maven.pkg.github.com/cicirello/Chips-n-Salsa</url>
+  <releases><enabled>true</enabled></releases>
+  <snapshots><enabled>true</enabled></snapshots>
+</repository>
+```
+
+__Step 3__: Run this via the command line.
+
+```
+$ mvn install
+```
+
 ## License
 
 The Chips-n-Salsa library is licensed under the [GNU General Public License 3.0](https://www.gnu.org/licenses/gpl-3.0.en.html).
