### 2020-07-28 3ad072f
Defined versioning scheme in README #24
fixes #24
