### 2020-07-28 0791c9a
Create CONTRIBUTING.md
Fixes #25
