### 2020-07-28 288f243
Update README.md
Added link to contribution guidelines.
