### 2020-07-28 8bbd03c
testcase for setup adjusted WSPT
