### 2020-07-28 7d7a98f
WSPT2 with setup adjustment
