### 2020-07-28 6f8f27c
Setup adjusted critical ratio heuristic
