### 2020-07-28 a9e437b
setup adjusted COVERT heuristic
