### 2020-07-28 d52e249
setup adjusted SPT
