+++ tests/org/cicirello/search/problems/scheduling/HeuristicTests.java
@@ -115,6 +115,28 @@ public class HeuristicTests {
 	}
 	
 	@Test
+	public void testSPTSetupAdjusted() {
+		double e = ShortestProcessingPlusSetupTime.MIN_H;
+		int highP = (int)Math.ceil(1 / e)*2;
+		int[] w =    { 1, 1, 1, 1, 0, 0, 0, 0, 2, 2, 2, 2, 1};
+		int[] p =    { 1, 2, 4, 8, 1, 2, 4, 8, 1, 2, 4, 8, highP};
+		double[] expected = { 1, 0.5, 0.25, 0.125, 1, 0.5, 0.25, 0.125, 1, 0.5, 0.25, 0.125, e};
+		PartialPermutation partial = new PartialPermutation(expected.length);
+		FakeProblemWeightsPTime problem = new FakeProblemWeightsPTime(w, p);
+		ShortestProcessingPlusSetupTime h = new ShortestProcessingPlusSetupTime(problem);
+		for (int j = 0; j < expected.length; j++) {
+			assertEquals(expected[j], h.h(partial, j, null), 1E-10);
+		}
+		
+		int[] ps =    { 0, 1, 3, 7, 0, 1, 3, 7, 0, 1, 3, 7, highP - 1};
+		problem = new FakeProblemWeightsPTime(w, ps, 0, 1);
+		h = new ShortestProcessingPlusSetupTime(problem);
+		for (int j = 0; j < expected.length; j++) {
+			assertEquals(expected[j], h.h(partial, j, null), 1E-10);
+		}
+	}
+	
+	@Test
 	public void testWSPT2() {
 		double e = WeightedShortestProcessingTimeLateOnly.MIN_H;
 		int highP = (int)Math.ceil(1 / e)*2;
