### 2020-07-28 332e972
smallest setup heuristic
