### 2020-07-28 a4ccd08
Smallest 2-job setup heuristic
