### 2020-07-28 7a9e9d8
smallest normalized setup
