+++ src/org/cicirello/search/problems/scheduling/SchedulingHeuristic.java
@@ -99,6 +99,23 @@ abstract class SchedulingHeuristic implements ConstructiveHeuristic {
 	/*
 	 * package-private rather than private to enable test case access
 	 */
+	int sumOfSetupTimes() {
+		if (!HAS_SETUPS) return 0;
+		int total = 0;
+		int n = data.numberOfJobs();
+		for (int i = 0; i <= n; i++) {
+			for (int j = 0; j < n; j++) {
+				if (i != j) {
+					total += data.getSetupTime(i, j);
+				}
+			}
+		}
+		return total;
+	}
+	
+	/*
+	 * package-private rather than private to enable test case access
+	 */
 	class IncrementalTimeCalculator implements IncrementalEvaluation {
 		
 		private int currentTime;
