### 2020-07-29 ce474af
changed getSetupTime method
