### 2020-07-29 6669804
fixed setup time summation
