### 2020-07-29 ccb894c
ATCS heuristic
