+++ src/org/cicirello/search/problems/scheduling/ATCS.java
@@ -1,0 +1,225 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.cicirello.search.problems.scheduling;
+
+import org.cicirello.permutations.Permutation;
+import org.cicirello.search.ss.IncrementalEvaluation;
+import org.cicirello.search.ss.PartialPermutation;
+
+/**
+ * <p>This is an implementation of the ATCS (Apparent Tardiness Cost with Setups)
+ * heuristic.   
+ * ATCS is defined as:
+ * h(j) = (w[j]/p[j]) exp( -max(0,d[j] - T - p[j]) / (k<sub>1</sub> p&#772;) ) 
+ * exp( -s[i][j] / (k<sub>2</sub> s&#772;)),
+ * where w[j] is the weight of job j, p[j] is its processing time,
+ * d[j] is the job's due date, T is the
+ * current time, and s[i][j] is any setup time of the job if it follows job i.  
+ * The k<sub>1</sub> and k<sub>2</sub> are parameters that can be tuned based on 
+ * problem instance characteristics, p&#772; is the average processing
+ * time of remaining unscheduled jobs, and s&#772; is the average setup time
+ * of the remaining unscheduled jobs.  The authors of the ATCS heuristic
+ * simply computed p&#772; and s&#772; once at the start, while our implementation
+ * updates these along the way as jobs are scheduled.</p>  
+ *
+ * <p>The constant {@link #MIN_H} defines the minimum value
+ * the heuristic will return, preventing h(j)=0 in support of stochastic
+ * sampling algorithms for which h(j)=0 is problematic.  This implementation 
+ * returns max( {@link #MIN_H}, h(j)), where {@link #MIN_H}
+ * is a small non-zero value.</p>  
+ *
+ * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
+ * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
+ * @version 7.29.2020
+ */
+public final class ATCS extends WeightedShortestProcessingTime {
+	
+	private final double k1;
+	private final double k2;
+	private final int pSum;
+	private final int sSum;
+	
+	/**
+	 * Constructs an ATCS heuristic.
+	 * @param problem The instance of a scheduling problem that is
+	 * the target of the heuristic.
+	 * @param k1 A parameter to the heuristic, which must be positive.
+	 * @param k2 A parameter to the heuristic, which must be positive.	 
[…]
