+++ tests/org/cicirello/search/problems/scheduling/HeuristicTests.java
@@ -83,12 +83,11 @@ public class HeuristicTests {
 		double e = SmallestSetup.MIN_H;
 		int highS = (int)Math.ceil(1 / e)*2;
 		int[][] s = {
-			{highS, highS, highS, highS, highS},
+			{0, highS, highS, highS, highS},
 			{1, 1, 7, 3, 1},
-			{1, 1, 1, 1, 1},
-			{1, 1, 1, 1, 1},
-			{1, 1, 1, 1, 1},
-			{0, 1, 3, 7, 15}
+			{1, 1, 3, 1, 1},
+			{1, 1, 1, 7, 1},
+			{1, 1, 1, 1, 15}
 		};
 		int[] w =    { 7, 8, 2, 10, 4};
 		int[] p =    { 2, 5, 9, 2, 10};
@@ -116,11 +115,10 @@ public class HeuristicTests {
 		int highS = (int)Math.ceil(1 / e)*2;
 		int[][] s = {
 			{0, 3, 2, 4, highS},
-			{1, 0, 2, 4, 1},
-			{1, 3, 0, 4, 1},
-			{1, 3, 1, 0, 1},
-			{1, 3, 1, 4, 0},
-			{0, 3, 9, 999, 999} // don't test last 2 jobs
+			{1, 3, 2, 4, 1},
+			{1, 3, 9, 4, 1},
+			{1, 3, 1, 999, 1},
+			{1, 3, 1, 4, 999} // don't test last 2 jobs
 		};
 		int[] w =    { 7, 8, 2, 10, 4};
 		int[] p =    { 2, 5, 9, 2, 10};
@@ -138,12 +136,11 @@ public class HeuristicTests {
 		double e = SmallestTwoJobSetup.MIN_H;
 		int highS = (int)Math.ceil(1 / e)*2;
 		int[][] s = {
-			{highS, highS, highS, 0, highS},
-			{1, 1, 7, 3, 1},
-			{9, 9, 9, 0, 9},
-			{7, 1, 7, 7, 7},
-			{5, 5, 2, 5, 5},
-			{0, 0, 3, 6, 13}
+			{0, highS, highS, 0, highS},
+			{1, 0, 7, 3, 1},
+			{9, 9, 3, 0, 9},
+			{7, 1, 7, 6, 7},
+			{5, 5, 2, 5, 13}
 		};
 		int[] w =    { 7, 8, 2, 10, 4};
 		int[] p =    { 2, 5, 9, 2, 10};
@@ -690,6 +687,98 @@ public class HeuristicTests {
 	}
 	
 	@Test
+	public void testATCS0() {
+		double e = ATCS.MIN_H;
+		int highP = (int)Math.ceil(1 / e)*2;
+		int[] w =    { 1, 1, 1, 1, 1, 0, 0, 0, 0, 2, 2, 2, 2, 1};
+		int[] p =    { 1, 1, 2, 4, 8, 1, 2, 4, 8, 1, 2, 4, 8, highP};
+		double[] expected0 = { 999, 1, 0.5, 0.25, 0.125, e, e, e, e, 2, 1, 0.5, 0.25, e};
+		double[] slack = new double[p.length];
+		double pAve = 0;
+		for (int i = 1; i < p.length; i++) {
+			slack[i] = 20-p[i]-p[0];
+			if (slack[i] < 0) slack[i] = 0;
+			pAve += p[i];
+		}
+		pAve /= p.length - 1;
+		PartialPermutation partial = new PartialPermutation(expected0.length);
+		//Doesn't really matter: partial.extend(0);
+		// All late tests, k1=2, k2=7?shouldn't matter
+		FakeProblemWeightsPTime problem = new FakeProblemWeightsPTime(w, p, 0);
[…]
