+++ CHANGELOG.md
@@ -1,0 +1,54 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased] - 2020-07-30
+### Added
+
+* ATCS scheduling heuristic.
+
+### Changed
+
+* Improved representation of setup times for scheduling problems with sequence-dependent setups.
+* Refactored ATC and Montagne scheduling heuristics, and scheduling heuristic abstract base class for performance and style.
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+
+## [1.2.0]
+### Added
+* Several constructive heuristics for scheduling problems for use with stochastic sampling search algorithms.
+* Heuristics include: EDD, MST, SPT, WSPT, Smallest Setup, Smallest Two-Job Setup, Montagne, Weighted COVERT, Weighted Critical Ratio, ATC.
+* Heuristics also include variations of several of the above that adjust for setup times for scheduling problems with sequence-dependent setups.
+* Contributing guidelines.
+
+### Changed
+* Refactored scheduling heuristic classes to improve readability.
+* Some code optimizations within the scheduling heuristic classes.
+* Improvements to README.
+
+## [1.1.0]
+### Added
+* Constructive heuristics for single machine scheduling problems.
+* Prebuilt jars now included with release.
+
+## [1.0.2]
+### Changed
+* Minor non-functional changes related to code style.
+* Edits to JOSS paper submission.
+
+## [1.0.1]
+### Other
+* Added a workflow to the repo to automatically publish to the GitHub Package Registry on release. This release is simply to confirm that it works and to publish the package to the GPR.
+
+## [1.0.0]
+### First public release
+Chips-n-Salsa is a Java library of customizable, hybridizable, iterative, parallel, stochastic, and self-adaptive local search algorithms. The library includes implementations of several stochastic local search algorithms, including simulated annealing, hill climbers, as well as constructive search algorithms such as stochastic sampling. The library most extensively supports simulated annealing. It includes several classes for representing solutions to a variety of optimization problems. For example, the library includes a BitVector class that implements vectors of bits, as well as classes for representing solutions to problems where we are searching for an optimal vector of integers or reals. For each of the built-in representations, the library provides the most common mutation operators for generating random neighbors of candidate solutions. Additionally, the library provides extensive support for permutation optimization problems, including implementations of many different mutation operators for permutations, and utilizing the efficiently implemented Permutation class of the JavaPermutationTools (JPT) library.
+
+Chips-n-Salsa is customizable, making extensive use of Java's generic types, enabling using the library to optimize other types of representations beyond what is provided in the library. It is hybridizable, providing support for integrating multiple forms of local search (e.g., using a hill climber on a solution generated by simulated annealing), creating hybrid mutation operators (e.g., local search using multiple mutation operators), as well as support for running more than one type of search for the same problem concurrently using multiple threads as a form of algorithm portfolio. Chips-n-Salsa is iterative, with support for multistart metaheuristics, including implementations of several restart schedules for varying the run lengths across the restarts. It also supports parallel execution of multiple instances of the same, or different, stochastic local search algorithms for an instance of a problem to accelerate the search process. The library supports self-adaptive search in a variety of ways, such as including implementations of adaptive annealing schedules for simulated annealing, such as the Modified Lam schedule, implementations of the simpler annealing schedules but which self-tune the initial temperature and other parameters, and restart schedules that adapt to run length.
