+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * ATCS scheduling heuristic.
+* CHANGELOG.md
 
 ### Changed
 
