### 2020-07-30 05214b4
New example program
Example program: industrial scheduling problem, VBSS, hybrid of VBSS and hill climbing.
