+++ examples/org/cicirello/examples/chipsnsalsa/SchedulingWithVBSS.java
@@ -140,10 +140,10 @@ public class SchedulingWithVBSS {
 		// We'll now construct the instance of the VBSS search.
 		// We pass the search a ConstructiveHeuristic, in this case the ATCS heuristic.
 		// We also optionally pass an exponent for a polynomial bias function
-		// (the default if left out is linear bias, but we use quadratic bias in this
-		// example).  You are not limited to polynomial bias functions, as there is
+		// (the default if left out is linear bias).  You are not limited to 
+		// polynomial bias functions, as there is
 		// a constructor that enables passing a custom bias function.
-		ValueBiasedStochasticSampling vbss = new ValueBiasedStochasticSampling(heuristic, 2);
+		ValueBiasedStochasticSampling vbss = new ValueBiasedStochasticSampling(heuristic, 8);
 		
 		// In the example, we're also going to demonstrate a hybrid of VBSS
 		// and hill climbing.  We use a first descent hill climber in the example.
@@ -170,7 +170,7 @@ public class SchedulingWithVBSS {
 			// performance comparison difficult.  So that we can examine problem
 			// solving performance separately, we'll create an independent instance
 			// of VBSS, but we'll use the same parameters.
-			new InitializeBySimpleMetaheuristic<Permutation>(new ValueBiasedStochasticSampling(heuristic, 2))
+			new InitializeBySimpleMetaheuristic<Permutation>(new ValueBiasedStochasticSampling(heuristic, 8))
 		);
 		
 		// We'll also compare to the solution generated by the deterministic application
@@ -198,7 +198,7 @@ public class SchedulingWithVBSS {
 		System.out.println("that we are minimizing.");
 		System.out.println();
 		System.out.printf("%7s\t%10s\t%10s\n", "Samples", "VBSS", "hybrid");
-		for (int totalSamples = 1, executedSamples = 0; totalSamples <= 10000; totalSamples *= 10) {
+		for (int totalSamples = 1, executedSamples = 0; totalSamples <= 100; totalSamples *= 10) {
 			int samples = totalSamples - executedSamples;
 			// Run each of the searches the number of samples to reach desired total.
 			// We'll ignore the return value of a SolutionCostPair, which returns the best of
