### 2020-07-30 4274c02
Renamed ATCS to DynamicATCS
The old ATCS class wasn't as described by the original authors.  It includes dynamic updates to heuristic parameters.  Renamed class, and will now introduce ATCS as originally described as an ATCS class.
