### 2020-07-30 4ddb90e
Original ATCS heuristic
