+++ .github/workflows/api-webpage-sitemap.yml
@@ -1,0 +1,35 @@
+on:
+  push:
+    branches:
+      - development
+
+jobs:
+  sitemap_job:
+    runs-on: ubuntu-latest
+    name: Generate a sitemap
+    steps:
+    - name: Checkout the repo
+      uses: actions/checkout@v2
+      with:
+        fetch-depth: 0 
+    - name: Generate the sitemap
+      id: sitemap
+      uses: cicirello/generate-sitemap@master
+      with:
+        base-url-path: https://chips-n-salsa.cicirello.org/
+        path-to-root: docs
+    - name: Output stats
+      run: |
+        echo "sitemap-path = ${{ steps.sitemap.outputs.sitemap-path }}"
+        echo "url-count = ${{ steps.sitemap.outputs.url-count }}"
+        echo "excluded-count = ${{ steps.sitemap.outputs.excluded-count }}"
+    - name: Create Pull Request
+      uses: peter-evans/create-pull-request@v3.1.0
+      with:
+        title: "Automated sitemap update"
+        body: > 
+          Sitemap was updated by [generate-sitemap](https://github.com/cicirello/generate-sitemap) 
+          GitHub action. Automated pull-request generated by 
+          [create-pull-request](https://github.com/peter-evans/create-pull-request) GitHub action.
+        commit-message: "[generate-sitemap] [create-pull-request] automated change."
+        
