+++ .github/workflows/api-webpage-sitemap.yml
@@ -1,3 +1,5 @@
+name: Generate API sitemap
+
 on:
   push:
     branches:
