### 2020-08-04 cicirello/Chips-n-Salsa#37 @mbdemoraes
Unable to run examples
I could download and build the library using the makefile. However, when I tried to run "make examples", I got the following error:

```
java -cp "exbin;dist/chips-n-salsa-1.0-jar-with-dependencies.jar" org.cicirello.examples.chipsnsalsa.BitVectorExample
Error: Could not locate or load main class org.cicirello.examples.chipsnsalsa.BitVectorExample
Caused by: java.lang.ClassNotFoundException: org.cicirello.examples.chipsnsalsa.BitVectorExample
make: *** [Makefile:11: examples] Error 1
```
The same error appears when I try to run each example manually. I'm running the commands from the root repository.

cc openjournals/joss-reviews#2448
