### 2020-08-04 cicirello/Chips-n-Salsa#37 @cicirello
@mbdemoraes Are you sure that the build was successful?  Can you check if the exbin directory was created (which is where the class files for the examples will be)?  I get that error if I run `make examples` without having executed the build.  Also, `make examples` by itself doesn't execute the build (there isn't really a good way in make to have that target in make dependent upon the build).
