### 2020-08-05 cicirello/Chips-n-Salsa#37 @mbdemoraes
Hey @cicirello ! Yeah, the build was successful and exbin directory was created. It seems every class is in there (BitVectorExample, CustomIntegerCostProblemExample, etc.).

I tried to download the library again, on another directory and did again the build process. Also, I tried to do a `make clean` and rebuild (with success), but in both cases I have the same issue.
