### 2020-08-05 cicirello/Chips-n-Salsa#37 @cicirello
@mbdemoraes I was able to recreate the makefile bug on Ubuntu.  I noticed that I had a semicolon as delimiter rather than a colon in the classpath when running the examples, so it worked on Windows, but not on Linux (or probably any other OS other than Windows). I have the makefile checking the OS now so should work on any OS.  Can you confirm that it works for you now?  Don't want to close issue yet in case there is something else hiding in the Makefile.  It now works for me on both Ubuntu and Windows.
