### 2020-08-05 64a0551
Refactored hill climbing classes
Reduced code redundancy in hill climbing classes with an abstract base class.
