+++ src/org/cicirello/search/hc/FirstDescentHillClimber.java
@@ -22,15 +22,12 @@ package org.cicirello.search.hc;
 
 import org.cicirello.util.Copyable;
 import org.cicirello.search.ProgressTracker;
-import org.cicirello.search.SolutionCostPair;
-import org.cicirello.search.problems.Problem;
 import org.cicirello.search.problems.OptimizationProblem;
 import org.cicirello.search.problems.IntegerCostOptimizationProblem;
 import org.cicirello.search.operators.IterableMutationOperator;
-import org.cicirello.search.operators.MutationIterator;
 import org.cicirello.search.operators.Initializer;
-import org.cicirello.search.Metaheuristic;
-import org.cicirello.search.SimpleLocalMetaheuristic;
+import org.cicirello.search.SolutionCostPair;
+import org.cicirello.search.operators.MutationIterator;
 
 
 /**
@@ -54,17 +51,9 @@ import org.cicirello.search.SimpleLocalMetaheuristic;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 6.15.2020
+ * @version 8.5.2020
  */
-public final class FirstDescentHillClimber<T extends Copyable<T>> implements Metaheuristic<T>, SimpleLocalMetaheuristic<T> {
-	
-	private final OptimizationProblem<T> pOpt;
-	private final IntegerCostOptimizationProblem<T> pOptInt;
-	private final Initializer<T> initializer;
-	private ProgressTracker<T> tracker;
-	private final IterableMutationOperator<T> mutation;
-	private final OneClimb<T> climber;
-	private long neighborCount;
+public final class FirstDescentHillClimber<T extends Copyable<T>> extends AbstractHillClimber<T> {
 	
 	/**
 	 * Constructs a first descent hill climber object for real-valued optimization problem.
@@ -76,15 +65,7 @@ public final class FirstDescentHillClimber<T extends Copyable<T>> implements Met
 	 * @throws NullPointerException if any of the parameters are null.
 	 */
 	public FirstDescentHillClimber(OptimizationProblem<T> problem, IterableMutationOperator<T> mutation, Initializer<T> initializer, ProgressTracker<T> tracker) {
-		if (problem == null || mutation == null || initializer == null || tracker == null) {
-			throw new NullPointerException();
-		}
-		pOpt = problem;
-		pOptInt = null;
-		this.mutation = mutation;
-		this.initializer = initializer;
-		this.tracker = tracker;
-		climber = initClimberDouble();
+		super(problem, mutation, initializer, tracker);
 	}
 	
 	/**
@@ -97,15 +78,7 @@ public final class FirstDescentHillClimber<T extends Copyable<T>> implements Met
 	 * @throws NullPointerException if any of the parameters are null.
 	 */
 	public FirstDescentHillClimber(IntegerCostOptimizationProblem<T> problem, IterableMutationOperator<T> mutation, Initializer<T> initializer, ProgressTracker<T> tracker) {
-		if (problem == null || mutation == null || initializer == null || tracker == null) {
-			throw new NullPointerException();
-		}
-		pOptInt = (IntegerCostOptimizationProblem<T>)problem;
-		pOpt = null;
-		this.mutation = mutation;
-		this.initializer = initializer;
-		this.tracker = tracker;
-		climber = initClimberInt();
+		super(problem, mutation, initializer, tracker);
 	}
 	
 	/**
@@ -117,7 +90,7 @@ public final class FirstDescentHillClimber<T extends Copyable<T>> implements Met
 	 * @throws NullPointerException if any of the parameters are null.
 	 */
[…]
