+++ src/org/cicirello/search/hc/AbstractHillClimber.java
@@ -1,0 +1,196 @@
+/*
+ * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
+ * Copyright (C) 2002-2020  Vincent A. Cicirello
+ *
+ * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
+ * 
+ * Chips-n-Salsa is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Chips-n-Salsa is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+ 
+package org.cicirello.search.hc;
+
+import org.cicirello.util.Copyable;
+import org.cicirello.search.ProgressTracker;
+import org.cicirello.search.SolutionCostPair;
+import org.cicirello.search.problems.Problem;
+import org.cicirello.search.problems.OptimizationProblem;
+import org.cicirello.search.problems.IntegerCostOptimizationProblem;
+import org.cicirello.search.operators.IterableMutationOperator;
+import org.cicirello.search.operators.Initializer;
+import org.cicirello.search.Metaheuristic;
+import org.cicirello.search.SimpleLocalMetaheuristic;
+
+
+/**
+ * <p>This class serves as an abstract base class for the Hill Climbing
+ * implementations, including the common functionality.</p> 
+ *
+ * @param <T> The type of object under optimization.
+ *
+ * @since 1.0
+ *
+ * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
+ * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
+ * @version 8.5.2020
+ */
+abstract class AbstractHillClimber<T extends Copyable<T>> implements Metaheuristic<T>, SimpleLocalMetaheuristic<T> {
+	
+	final OptimizationProblem<T> pOpt;
+	final IntegerCostOptimizationProblem<T> pOptInt;
+	private final Initializer<T> initializer;
+	ProgressTracker<T> tracker;
+	final IterableMutationOperator<T> mutation;
+	private final OneClimb<T> climber;
+	long neighborCount;
+	
+	/**
+	 * Constructs a first descent hill climber object for real-valued optimization problem.
+	 * @param problem An instance of an optimization problem to solve.
+	 * @param mutation A mutation operator.
+	 * @param initializer The source of random initial states for each hill climb.
+	 * @param tracker A ProgressTracker object, which is used to keep track of the best
+	 * solution found during the run, the time when it was found, and other related data.
+	 * @throws NullPointerException if any of the parameters are null.
+	 */
+	AbstractHillClimber(OptimizationProblem<T> problem, IterableMutationOperator<T> mutation, Initializer<T> initializer, ProgressTracker<T> tracker) {
+		if (problem == null || mutation == null || initializer == null || tracker == null) {
+			throw new NullPointerException();
+		}
+		pOpt = problem;
+		pOptInt = null;
+		this.mutation = mutation;
+		this.initializer = initializer;
+		this.tracker = tracker;
+		climber = initClimberDouble();
+	}
+	
+	/**
+	 * Constructs a first descent hill climber object for integer-valued optimization problem.
+	 * @param problem An instance of an optimization problem to solve.
[…]

[…]
