### 2020-08-05 cicirello/Chips-n-Salsa#37 @mbdemoraes
@cicirello thank you for your modification! I forgot to say that I was using Lubuntu (an Ubuntu-based distribution), I'm sorry. But now it works fine.
