### 2020-08-06 6214f78
Update Makefile
Previous bug fix inadvertantly left off one of the example programs.  Added back to Makefile.
