### 2020-08-06 c2c107b
minor documentation edits
