### 2020-08-07 #2448 @galessiorob
@cicirello thanks for working out the install and run instructions with @mbdemoraes!

@mbdemoraes thank you so much for your review, the only [item to check-off](https://github.com/openjournals/joss-reviews/issues/2448#issue-652769899) is `Performance`, which based on the article you can check because there are no performance claims.

@nnadeau thank you too for your review, the only missing items from[ your checklist](https://github.com/openjournals/joss-reviews/issues/2448#issue-652769899) are `Installation instructions` and `Community guidelines`, mind checking those off if you're satisfied or giving notes on what is missing if so?

Almost there! 🎉
