### 2020-08-07 40fcf23
refactored VBSS
reduced code redundancy using abstract base class.
