+++ src/org/cicirello/search/ss/ValueBiasedStochasticSampling.java
@@ -21,10 +21,6 @@
 package org.cicirello.search.ss;
 
 import org.cicirello.permutations.Permutation;
-import org.cicirello.search.SimpleMetaheuristic;
-import org.cicirello.search.Metaheuristic;
-import org.cicirello.search.problems.Problem;
-import org.cicirello.search.problems.IntegerCostOptimizationProblem;
 import org.cicirello.search.SolutionCostPair;
 import org.cicirello.search.ProgressTracker;
 import java.util.concurrent.ThreadLocalRandom;
@@ -140,14 +136,12 @@ import java.util.concurrent.ThreadLocalRandom;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 7.9.2020
+ * @version 8.7.2020
  */
-public final class ValueBiasedStochasticSampling implements SimpleMetaheuristic<Permutation>, Metaheuristic<Permutation> {
+public final class ValueBiasedStochasticSampling extends AbstractStochasticSampler {
 	
-	private final VBSSSampler sampler;
-	private ProgressTracker<Permutation> tracker;
-	private int numGenerated;
 	private final BiasFunction bias;
+	private final ConstructiveHeuristic heuristic;
 	
 	/**
 	 * Constructs a ValueBiasedStochasticSampling search object.  A ProgressTracker 
@@ -215,83 +209,18 @@ public final class ValueBiasedStochasticSampling implements SimpleMetaheuristic<
 	 * @throws NullPointerException if heuristic or tracker is null
 	 */
 	public ValueBiasedStochasticSampling(ConstructiveHeuristic heuristic, BiasFunction bias, ProgressTracker<Permutation> tracker) {
-		if (heuristic == null || tracker == null) {
-			throw new NullPointerException();
-		}
-		this.tracker = tracker;
+		super(heuristic.getProblem(), tracker);
 		this.bias = bias;
-		// default: numGenerated = 0;
-		if (heuristic.getProblem() instanceof IntegerCostOptimizationProblem) {
-			sampler = new IntCost(heuristic);
-		} else {
-			sampler = new DoubleCost(heuristic);
-		}
+		this.heuristic = heuristic;
 	}
 	
 	/*
 	 * private for use by split method
 	 */
 	private ValueBiasedStochasticSampling(ValueBiasedStochasticSampling other) {
-		tracker = other.tracker;
+		super(other);
 		bias = other.bias;
-		if (other.sampler instanceof IntCost) {
-			sampler = new IntCost((IntCost)other.sampler);
-		} else {
-			sampler = new DoubleCost((DoubleCost)other.sampler);
-		}
-		// default: numGenerated = 0;
-	}
-	
-	@Override
-	public SolutionCostPair<Permutation> optimize() {
-		if (tracker.didFindBest() || tracker.isStopped()) return null;
-		numGenerated++;
-		return sampler.optimize();
-	}
-	
-	/**
-	 * <p>Generates multiple samples using Value Biased Stochastic Sampling
-	 * (VBSS).  Returns the best solution of the set of samples.</p>
-	 *
-	 * @param numSamples The number of samples of VBSS to perform.
-	 * @return The best solution of this set of samples, which may or may not be the 
-	 * same as the solution contained
-	 * in this search's {@link org.cicirello.search.ProgressTracker ProgressTracker}, 
-	 * which contains the best of all runs
-	 * across all calls to the various optimize methods.
-	 * Returns null if no runs executed, such as if the ProgressTracker already contains
-	 * the theoretical best solution.
-	 */
[…]
