### 2020-08-07 1acc9f5
refactored AcceptanceBandSampling
removed code redundancy using the new abstract base class
