+++ src/org/cicirello/search/ss/AcceptanceBandSampling.java
@@ -21,10 +21,6 @@
 package org.cicirello.search.ss;
 
 import org.cicirello.permutations.Permutation;
-import org.cicirello.search.SimpleMetaheuristic;
-import org.cicirello.search.Metaheuristic;
-import org.cicirello.search.problems.Problem;
-import org.cicirello.search.problems.IntegerCostOptimizationProblem;
 import org.cicirello.search.SolutionCostPair;
 import org.cicirello.search.ProgressTracker;
 import org.cicirello.math.rand.RandomIndexer;
@@ -109,14 +105,12 @@ import org.cicirello.math.rand.RandomIndexer;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 7.10.2020
+ * @version 8.7.2020
  */
-public final class AcceptanceBandSampling implements SimpleMetaheuristic<Permutation>, Metaheuristic<Permutation> {
+public final class AcceptanceBandSampling extends AbstractStochasticSampler {
 	
-	private final Sampler sampler;
+	private final ConstructiveHeuristic heuristic;
 	private final double acceptancePercentage;
-	private ProgressTracker<Permutation> tracker;
-	private int numGenerated;
 	
 	/**
 	 * Constructs an AcceptanceBandSampling search object.  
@@ -178,86 +172,21 @@ public final class AcceptanceBandSampling implements SimpleMetaheuristic<Permuta
 	 * @throws IllegalArgumentException if beta is less than 0.0 or greater than 1.0.
 	 */
 	public AcceptanceBandSampling(ConstructiveHeuristic heuristic, double beta, ProgressTracker<Permutation> tracker) {
-		if (heuristic == null || tracker == null) {
-			throw new NullPointerException();
-		}
+		super(heuristic.getProblem(), tracker);
+		this.heuristic = heuristic;
 		if (beta < 0.0 || beta > 1.0) {
 			throw new IllegalArgumentException("beta must be in the interval: [0.0, 1.0].");
 		}
-		this.tracker = tracker;
 		acceptancePercentage = 1.0 - beta;
-		if (heuristic.getProblem() instanceof IntegerCostOptimizationProblem) {
-			sampler = new IntCost(heuristic);
-		} else {
-			sampler = new DoubleCost(heuristic);
-		}
-		// default: numGenerated = 0;
 	}
 	
 	/*
 	 * private for use by split method
 	 */
 	private AcceptanceBandSampling(AcceptanceBandSampling other) {
-		tracker = other.tracker;
+		super(other);
+		heuristic = other.heuristic;
 		acceptancePercentage = other.acceptancePercentage;
-		if (other.sampler instanceof IntCost) {
-			sampler = new IntCost((IntCost)other.sampler);
-		} else {
-			sampler = new DoubleCost((DoubleCost)other.sampler);
-		}
-		// default: numGenerated = 0;
-	}
-	
-	@Override
-	public SolutionCostPair<Permutation> optimize() {
-		if (tracker.didFindBest() || tracker.isStopped()) return null;
-		numGenerated++;
-		return sampler.optimize();
-	}
-	
-	/**
-	 * <p>Generates multiple samples using Acceptance Band Sampling.  
-	 * Returns the best solution of the set of samples.</p>
-	 *
-	 * @param numSamples The number of samples to perform.
-	 * @return The best solution of this set of samples, which may or may not be the 
-	 * same as the solution contained
-	 * in this search's {@link org.cicirello.search.ProgressTracker ProgressTracker}, 
-	 * which contains the best of all runs
[…]
