### 2020-08-11 #2448 @galessiorob
@cicirello we're almost ready to publish! Could you please deposit your software in Zenodo and post the DOI here?

After that, we can set the final version and accept the paper 🎉

@mbdemoraes and @nnadeau thank you both so much for your time and expertise ✨
