### 2020-08-11 #2448 @cicirello
@galessiorob the Zenodo doi for all versions is https://doi.org/10.5281/zenodo.3959773 . I think that is probably the one you want. It looks like zenodo also generates a doi for each release, so if you want the one that goes to a specific release then it is https://doi.org/10.5281/zenodo.3974614 .
