### 2020-08-12 2d7714f
Update CHANGELOG.md
