+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2020-08-07
+## [Unreleased] - 2020-08-11
 ### Added
+* .zenodo.json to provide project metadata to Zenodo during releases
 
 ### Changed
 * Refactored HBSS, VBSS, and AcceptanceBandSampling to use a common abstract base class to remove code redundancy.
