### 2020-08-12 1df8400
refactored AbstractStochasticSampler
