+++ src/org/cicirello/search/ss/ValueBiasedStochasticSampling.java
@@ -136,9 +136,9 @@ import java.util.concurrent.ThreadLocalRandom;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 8.7.2020
+ * @version 8.12.2020
  */
-public final class ValueBiasedStochasticSampling extends AbstractStochasticSampler {
+public final class ValueBiasedStochasticSampling extends AbstractStochasticSampler<Permutation> {
 	
 	private final BiasFunction bias;
 	private final ConstructiveHeuristic heuristic;
@@ -292,7 +292,7 @@ public final class ValueBiasedStochasticSampling extends AbstractStochasticSampl
 		else return select(values, mid+1, last, u);
 	}
 	
-	Sampler initSamplerInt() {
+	Sampler<Permutation> initSamplerInt() {
 		return () -> {
 			IncrementalEvaluation incEval = heuristic.createIncrementalEvaluation();
 			int n = heuristic.completePermutationLength();
@@ -324,7 +324,7 @@ public final class ValueBiasedStochasticSampling extends AbstractStochasticSampl
 		};
 	}
 	
-	Sampler initSamplerDouble() {
+	Sampler<Permutation> initSamplerDouble() {
 		return () -> {
 			IncrementalEvaluation incEval = heuristic.createIncrementalEvaluation();
 			int n = heuristic.completePermutationLength();
