+++ src/org/cicirello/search/ss/AcceptanceBandSampling.java
@@ -105,9 +105,9 @@ import org.cicirello.math.rand.RandomIndexer;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 8.7.2020
+ * @version 8.12.2020
  */
-public final class AcceptanceBandSampling extends AbstractStochasticSampler {
+public final class AcceptanceBandSampling extends AbstractStochasticSampler<Permutation> {
 	
 	private final ConstructiveHeuristic heuristic;
 	private final double acceptancePercentage;
@@ -209,7 +209,7 @@ public final class AcceptanceBandSampling extends AbstractStochasticSampler {
 		return equivalents[RandomIndexer.nextInt(n)];
 	}
 	
-	Sampler initSamplerInt() {
+	Sampler<Permutation> initSamplerInt() {
 		return () -> {
 			IncrementalEvaluation incEval = heuristic.createIncrementalEvaluation();
 			int n = heuristic.completePermutationLength();
@@ -242,7 +242,7 @@ public final class AcceptanceBandSampling extends AbstractStochasticSampler {
 		};
 	}
 	
-	Sampler initSamplerDouble() {
+	Sampler<Permutation> initSamplerDouble() {
 		return () -> {
 			IncrementalEvaluation incEval = heuristic.createIncrementalEvaluation();
 			int n = heuristic.completePermutationLength();
