+++ src/org/cicirello/search/ss/AbstractStochasticSampler.java
@@ -20,7 +20,6 @@
  
 package org.cicirello.search.ss;
 
-import org.cicirello.permutations.Permutation;
 import org.cicirello.search.SimpleMetaheuristic;
 import org.cicirello.search.Metaheuristic;
 import org.cicirello.search.problems.IntegerCostOptimizationProblem;
@@ -28,7 +27,7 @@ import org.cicirello.search.problems.OptimizationProblem;
 import org.cicirello.search.problems.Problem;
 import org.cicirello.search.ProgressTracker;
 import org.cicirello.search.SolutionCostPair;
-
+import org.cicirello.util.Copyable;
 
 /**
  * <p>This class serves as an abstract base class for the stochastic
@@ -38,14 +37,14 @@ import org.cicirello.search.SolutionCostPair;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 8.7.2020
+ * @version 8.12.2020
  */
-abstract class AbstractStochasticSampler implements SimpleMetaheuristic<Permutation>, Metaheuristic<Permutation> {
+abstract class AbstractStochasticSampler<T extends Copyable<T>> implements SimpleMetaheuristic<T>, Metaheuristic<T> {
 	
-	final OptimizationProblem<Permutation> pOpt;
-	final IntegerCostOptimizationProblem<Permutation> pOptInt;
-	ProgressTracker<Permutation> tracker;
-	private final Sampler sampler;
+	final OptimizationProblem<T> pOpt;
+	final IntegerCostOptimizationProblem<T> pOptInt;
+	ProgressTracker<T> tracker;
+	private final Sampler<T> sampler;
 	private int numGenerated;
 	
 	/**
@@ -54,18 +53,18 @@ abstract class AbstractStochasticSampler implements SimpleMetaheuristic<Permutat
 	 * @param tracker A ProgressTracker
 	 * @throws NullPointerException if problem or tracker is null.
 	 */
-	AbstractStochasticSampler(Problem<Permutation> problem, ProgressTracker<Permutation> tracker) {
+	AbstractStochasticSampler(Problem<T> problem, ProgressTracker<T> tracker) {
 		if (problem == null || tracker == null) {
 			throw new NullPointerException();
 		}
 		this.tracker = tracker;
 		// default: numGenerated = 0;
 		if (problem instanceof IntegerCostOptimizationProblem) {
-			pOptInt = (IntegerCostOptimizationProblem<Permutation>)problem;
+			pOptInt = (IntegerCostOptimizationProblem<T>)problem;
 			pOpt = null;
 			sampler = initSamplerInt();
 		} else {
-			pOpt = (OptimizationProblem<Permutation>)problem;
+			pOpt = (OptimizationProblem<T>)problem;
 			pOptInt = null;
 			sampler = initSamplerDouble();
 		}
@@ -75,7 +74,7 @@ abstract class AbstractStochasticSampler implements SimpleMetaheuristic<Permutat
 	 * package-private copy constructor in support of the split method.
 	 * note: copies references to thread-safe components, and splits potentially non-threadsafe components 
 	 */
-	AbstractStochasticSampler(AbstractStochasticSampler other) {
+	AbstractStochasticSampler(AbstractStochasticSampler<T> other) {
 		// these are threadsafe, so just copy references
 		pOpt = other.pOpt;
 		pOptInt = other.pOptInt;
@@ -90,7 +89,7 @@ abstract class AbstractStochasticSampler implements SimpleMetaheuristic<Permutat
 	}
 	
 	@Override
[…]
