### 2020-08-12 3619e4b
refactored IterativeSampling
Refactored IterativeSampling to use the abstract base class AbstractStochasticSampler to remove redundant code.
