+++ src/org/cicirello/search/ss/IterativeSampling.java
@@ -23,12 +23,9 @@ package org.cicirello.search.ss;
 import org.cicirello.util.Copyable;
 import org.cicirello.search.ProgressTracker;
 import org.cicirello.search.SolutionCostPair;
-import org.cicirello.search.problems.Problem;
 import org.cicirello.search.problems.OptimizationProblem;
 import org.cicirello.search.problems.IntegerCostOptimizationProblem;
 import org.cicirello.search.operators.Initializer;
-import org.cicirello.search.Metaheuristic;
-import org.cicirello.search.SimpleMetaheuristic;
 
 /**
  * <p>Iterative sampling is the simplest possible form of a stochastic sampling search.
@@ -49,16 +46,11 @@ import org.cicirello.search.SimpleMetaheuristic;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 6.15.2020
+ * @version 8.12.2020
  */
-public final class IterativeSampling<T extends Copyable<T>> implements Metaheuristic<T>, SimpleMetaheuristic<T> {
+public final class IterativeSampling<T extends Copyable<T>> extends AbstractStochasticSampler<T> {
 	
-	private final OptimizationProblem<T> pOpt;
-	private final IntegerCostOptimizationProblem<T> pOptInt;
 	private final Initializer<T> initializer;
-	private ProgressTracker<T> tracker;
-	private long totalNumSamples;
-	private final OneSample<T> sampler;
 	
 	/**
 	 * Constructs an iterative sampling search for a real-valued optimization problem.
@@ -69,15 +61,11 @@ public final class IterativeSampling<T extends Copyable<T>> implements Metaheuri
 	 * @throws NullPointerException if any of the parameters are null.
 	 */
 	public IterativeSampling(OptimizationProblem<T> problem, Initializer<T> initializer, ProgressTracker<T> tracker) {
-		if (problem == null || initializer == null || tracker == null) {
+		super(problem, tracker);
+		if (initializer == null) {
 			throw new NullPointerException();
 		}
-		pOpt = problem;
-		pOptInt = null;
 		this.initializer = initializer;
-		this.tracker = tracker;
-		// Deliberately using default: totalNumSamples = 0;
-		sampler = initSamplerDouble();
 	}
 	
 	/**
@@ -89,15 +77,11 @@ public final class IterativeSampling<T extends Copyable<T>> implements Metaheuri
 	 * @throws NullPointerException if any of the parameters are null.
 	 */
 	public IterativeSampling(IntegerCostOptimizationProblem<T> problem, Initializer<T> initializer, ProgressTracker<T> tracker) {
-		if (problem == null || initializer == null || tracker == null) {
+		super(problem, tracker);
+		if (initializer == null) {
 			throw new NullPointerException();
 		}
-		pOptInt = problem;
-		pOpt = null;
 		this.initializer = initializer;
-		this.tracker = tracker;
-		// Deliberately using default: totalNumSamples = 0;
-		sampler = initSamplerInt();
 	}
 	
 	/**
@@ -127,79 +111,9 @@ public final class IterativeSampling<T extends Copyable<T>> implements Metaheuri
 	 * note: copies references to thread-safe components, and splits potentially non-threadsafe components 
 	 */
 	private IterativeSampling(IterativeSampling<T> other) {
-		// these are threadsafe, so just copy references
-		pOpt = other.pOpt;
-		pOptInt = other.pOptInt;
-		
-		// this one must be shared so just copy reference.
[…]
