### 2020-08-12 e92b9b6
check for best possible solution
Added missing check if a solution is the theoretical best for problem (e.g., to terminate search rather than proceeding needlessly).
