+++ src/org/cicirello/search/ss/ValueBiasedStochasticSampling.java
@@ -319,6 +319,9 @@ public final class ValueBiasedStochasticSampling extends AbstractStochasticSampl
 			int cost = solution.getCost();
 			if (cost < tracker.getCost()) {
 				tracker.update(cost, complete);
+				if (cost == pOptInt.minCost()) {
+					tracker.setFoundBest();
+				}
 			}
 			return solution;
 		};
@@ -351,6 +354,9 @@ public final class ValueBiasedStochasticSampling extends AbstractStochasticSampl
 			double cost = solution.getCostDouble();
 			if (cost < tracker.getCostDouble()) {
 				tracker.update(cost, complete);
+				if (cost == pOpt.minCost()) {
+					tracker.setFoundBest();
+				}
 			}
 			return solution;
 		};
