+++ src/org/cicirello/search/ss/AcceptanceBandSampling.java
@@ -237,6 +237,9 @@ public final class AcceptanceBandSampling extends AbstractStochasticSampler<Perm
 			int cost = solution.getCost();
 			if (cost < tracker.getCost()) {
 				tracker.update(cost, complete);
+				if (cost == pOptInt.minCost()) {
+					tracker.setFoundBest();
+				}
 			}
 			return solution;
 		};
@@ -270,6 +273,9 @@ public final class AcceptanceBandSampling extends AbstractStochasticSampler<Perm
 			double cost = solution.getCostDouble();
 			if (cost < tracker.getCostDouble()) {
 				tracker.update(cost, complete);
+				if (cost == pOpt.minCost()) {
+					tracker.setFoundBest();
+				}
 			}
 			return solution;
 		};
