### 2020-08-12 f62b902
Update CHANGELOG.md
