+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * .zenodo.json to provide project metadata to Zenodo during releases
 
 ### Changed
-* Refactored HBSS, VBSS, and AcceptanceBandSampling to use a common abstract base class to remove code redundancy.
+* Refactored HBSS, VBSS, AcceptanceBandSampling, and IterativeSampling to use a common abstract base class to remove code redundancy.
 
 ### Deprecated
 
