### 2020-08-12 a002e92
Refactored HeuristicPermutationGenerator
redeuced code redundancy
