+++ src/org/cicirello/search/ss/HeuristicPermutationGenerator.java
@@ -24,6 +24,7 @@ import org.cicirello.permutations.Permutation;
 import org.cicirello.search.SimpleMetaheuristic;
 import org.cicirello.search.problems.Problem;
 import org.cicirello.search.problems.IntegerCostOptimizationProblem;
+import org.cicirello.search.problems.OptimizationProblem;
 import org.cicirello.search.SolutionCostPair;
 import org.cicirello.search.ProgressTracker;
 
@@ -54,11 +55,14 @@ import org.cicirello.search.ProgressTracker;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 7.7.2020
+ * @version 8.12.2020
  */
 public final class HeuristicPermutationGenerator implements SimpleMetaheuristic<Permutation> {
 	
+	private final OptimizationProblem<Permutation> pOpt;
+	private final IntegerCostOptimizationProblem<Permutation> pOptInt;
 	private final HeuristicGenerator generator;
+	private final ConstructiveHeuristic heuristic;
 	private ProgressTracker<Permutation> tracker;
 	private int numGenerated;
 	
@@ -85,11 +89,17 @@ public final class HeuristicPermutationGenerator implements SimpleMetaheuristic<
 			throw new NullPointerException();
 		}
 		this.tracker = tracker;
+		this.heuristic = heuristic;
 		// default: numGenerated = 0;
+		Problem<Permutation> problem = heuristic.getProblem();
 		if (heuristic.getProblem() instanceof IntegerCostOptimizationProblem) {
-			generator = new IntCost(heuristic);
+			pOptInt = (IntegerCostOptimizationProblem<Permutation>)problem;
+			pOpt = null;
+			generator = initGeneratorInt();
 		} else {
-			generator = new DoubleCost(heuristic);
+			pOpt = (OptimizationProblem<Permutation>)problem;
+			pOptInt = null;
+			generator = initGeneratorDouble();
 		}
 	}
 	
@@ -97,17 +107,26 @@ public final class HeuristicPermutationGenerator implements SimpleMetaheuristic<
 	 * private for use by split method
 	 */
 	private HeuristicPermutationGenerator(HeuristicPermutationGenerator other) {
+		// these are threadsafe, so just copy references
+		pOpt = other.pOpt;
+		pOptInt = other.pOptInt;
+		heuristic = other.heuristic;
+		
+		// this one must be shared.
 		tracker = other.tracker;
-		if (other.generator instanceof IntCost) {
-			generator = new IntCost((IntCost)other.generator);
-		} else {
-			generator = new DoubleCost((DoubleCost)other.generator);
-		}
+		
+		// not threadsafe
+		generator = pOptInt != null ? initGeneratorInt() : initGeneratorDouble();
+
 		// default: numGenerated = 0;
 	}
 	
 	@Override
 	public SolutionCostPair<Permutation> optimize() {
+		if (tracker.isStopped() || tracker.didFindBest()) {
+			return null;
+		}
+		numGenerated++;
 		return generator.optimize();
 	}
 	
@@ -128,7 +147,7 @@ public final class HeuristicPermutationGenerator implements SimpleMetaheuristic<
 	
 	@Override
 	public Problem<Permutation> getProblem() {
-		return generator.getProblem();
+		return (pOptInt != null) ? pOptInt : pOpt;
 	}
 	
 	@Override
@@ -138,32 +157,10 @@ public final class HeuristicPermutationGenerator implements SimpleMetaheuristic<
 	
 	private interface HeuristicGenerator {
 		SolutionCostPair<Permutation> optimize();
[…]
