### 2020-08-12 bae6bc2
Update CHANGELOG.md
