### 2020-08-12 cd34373
simplified stochastic samplers
