+++ src/org/cicirello/search/ss/HeuristicBiasedStochasticSampling.java
@@ -370,77 +370,31 @@ public final class HeuristicBiasedStochasticSampling extends AbstractStochasticS
 		return pivot;
 	}
 	
-	Sampler<Permutation> initSamplerInt() {
-		return () -> {
-			IncrementalEvaluation incEval = heuristic.createIncrementalEvaluation();
-			int n = heuristic.completePermutationLength();
-			PartialPermutation p = new PartialPermutation(n);
-			double[] v = new double[n];
-			int[] extensions = new int[n];
-			ThreadLocalRandom r = ThreadLocalRandom.current();
-			while (!p.isComplete()) {
-				int k = p.numExtensions();
-				if (k==1) {
-					incEval.extend(p, p.getExtension(0));
-					p.extend(0);
-				} else {
-					int chosenRank = 1 + select(biases, k, r.nextDouble(biases[k-1]));
-					for (int i = 0; i < k; i++) {
-						v[i] = heuristic.h(p, p.getExtension(i), incEval);
-						extensions[i] = i;
-					}
-					int which = randomizedSelect(extensions, v, k, chosenRank);
-					incEval.extend(p, p.getExtension(which));
-					p.extend(which);
+	@Override
+	SolutionCostPair<Permutation> sample() {
+		IncrementalEvaluation incEval = heuristic.createIncrementalEvaluation();
+		int n = heuristic.completePermutationLength();
+		PartialPermutation p = new PartialPermutation(n);
+		double[] v = new double[n];
+		int[] extensions = new int[n];
+		ThreadLocalRandom r = ThreadLocalRandom.current();
+		while (!p.isComplete()) {
+			int k = p.numExtensions();
+			if (k==1) {
+				incEval.extend(p, p.getExtension(0));
+				p.extend(0);
+			} else {
+				int chosenRank = 1 + select(biases, k, r.nextDouble(biases[k-1]));
+				for (int i = 0; i < k; i++) {
+					v[i] = heuristic.h(p, p.getExtension(i), incEval);
+					extensions[i] = i;
 				}
+				int which = randomizedSelect(extensions, v, k, chosenRank);
+				incEval.extend(p, p.getExtension(which));
+				p.extend(which);
 			}
-			Permutation complete = p.toComplete();
-			SolutionCostPair<Permutation> solution = pOptInt.getSolutionCostPair(complete);
-			int cost = solution.getCost();
-			if (cost < tracker.getCost()) {
-				tracker.update(cost, complete);
-				if (cost == pOptInt.minCost()) {
-					tracker.setFoundBest();
-				}
-			}
-			return solution;
-		};
-	}
-	
-	Sampler<Permutation> initSamplerDouble() {
-		return () -> {
-			IncrementalEvaluation incEval = heuristic.createIncrementalEvaluation();
-			int n = heuristic.completePermutationLength();
-			PartialPermutation p = new PartialPermutation(n);
-			double[] v = new double[n];
-			int[] extensions = new int[n];
-			ThreadLocalRandom r = ThreadLocalRandom.current();
-			while (!p.isComplete()) {
-				int k = p.numExtensions();
-				if (k==1) {
-					incEval.extend(p, p.getExtension(0));
-					p.extend(0);
-				} else {
-					int chosenRank = 1 + select(biases, k, r.nextDouble(biases[k-1]));
-					for (int i = 0; i < k; i++) {
-						v[i] = heuristic.h(p, p.getExtension(i), incEval);
-						extensions[i] = i;
-					}
-					int which = randomizedSelect(extensions, v, k, chosenRank);
-					incEval.extend(p, p.getExtension(which));
-					p.extend(which);
-				}
-			}
-			Permutation complete = p.toComplete();
[…]
