### 2020-08-12 ddddb3f
further refactoring
