### 2020-08-12 1bd2675
refactrored SimulatedAnnealing
