### 2020-08-12 a65e989
Update CHANGELOG.md
