### 2020-08-13 adc0460
further refactoring
