+++ src/org/cicirello/search/sa/SimulatedAnnealing.java
@@ -89,10 +89,11 @@ import org.cicirello.search.SimpleLocalMetaheuristic;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 6.15.2020
+ * @version 8.12.2020
  */
 public class SimulatedAnnealing<T extends Copyable<T>> implements SingleSolutionMetaheuristic<T> {
 	
+	private final SimpleLocalMetaheuristic<T> hc;
 	private final IntegerCostOptimizationProblem<T> pOptInt;
 	private final OptimizationProblem<T> pOpt;
 	private final Initializer<T> initializer;
@@ -106,23 +107,17 @@ public class SimulatedAnnealing<T extends Copyable<T>> implements SingleSolution
 	 * internal constructor
 	 */
 	private SimulatedAnnealing(IntegerCostOptimizationProblem<T> problem, UndoableMutationOperator<T> mutation, Initializer<T> initializer, AnnealingSchedule anneal, ProgressTracker<T> tracker) {
-		if (problem == null || mutation == null || anneal == null || initializer == null || tracker == null) {
-			throw new NullPointerException();
-		}
-		this.initializer = initializer;
-		this.mutation = mutation;
-		this.anneal = anneal;
-		this.tracker = tracker;
-		pOptInt = problem;
-		pOpt = null;
-		// default on purpose: elapsedEvals = 0;
-		sr = initSingleRunInt();
+		this(problem, mutation, initializer, anneal, tracker, null);
 	}
 	
 	/*
 	 * internal constructor
 	 */
 	private SimulatedAnnealing(OptimizationProblem<T> problem, UndoableMutationOperator<T> mutation, Initializer<T> initializer, AnnealingSchedule anneal, ProgressTracker<T> tracker) {
+		this(problem, mutation, initializer, anneal, tracker, null);
+	}
+	
+	private SimulatedAnnealing(OptimizationProblem<T> problem, UndoableMutationOperator<T> mutation, Initializer<T> initializer, AnnealingSchedule anneal, ProgressTracker<T> tracker, SimpleLocalMetaheuristic<T> hc) {
 		if (problem == null || mutation == null || anneal == null || initializer == null || tracker == null) {
 			throw new NullPointerException();
 		}
@@ -134,6 +129,40 @@ public class SimulatedAnnealing<T extends Copyable<T>> implements SingleSolution
 		pOptInt = null;
 		// default on purpose: elapsedEvals = 0;
 		sr = initSingleRunDouble();
+		
+		this.hc = hc;
+		if (hc != null) {
+			if (problem != hc.getProblem()) {
+				throw new IllegalArgumentException("hc must be configured with the same problem.");
+			}
+			if (hc.getProgressTracker() != tracker) {
+				hc.setProgressTracker(tracker);
+			}
+		}
+	}
+	
+	private SimulatedAnnealing(IntegerCostOptimizationProblem<T> problem, UndoableMutationOperator<T> mutation, Initializer<T> initializer, AnnealingSchedule anneal, ProgressTracker<T> tracker, SimpleLocalMetaheuristic<T> hc) {
+		if (problem == null || mutation == null || anneal == null || initializer == null || tracker == null) {
+			throw new NullPointerException();
+		}
+		this.initializer = initializer;
+		this.mutation = mutation;
+		this.anneal = anneal;
+		this.tracker = tracker;
+		pOptInt = problem;
+		pOpt = null;
+		// default on purpose: elapsedEvals = 0;
[…]
