### 2020-08-13 3b3c502
refactored constructors
Refactored the constructors of the SimulatedAnnealing class, to replace the factory methods.  The factory method previously supported a more complex hierarchy of internal classes that have since been simplified.   Adjusted the example programs to use the new simplified constructors of the SimulatedAnnealing class.
