+++ examples/org/cicirello/examples/chipsnsalsa/PostHillclimbExample.java
@@ -89,11 +89,11 @@ public class PostHillclimbExample {
 		// mutation operator, the initializer, and the hill climber.  
 		// This example uses the default
 		// annealing schedule, the Modified Lam schedule.  There are other
-		// factory methods that allow specifying a different annealing schedule.
+		// constructors that allow specifying a different annealing schedule.
 		// The library uses generics so that simulated annealing can easily be applied
 		// to optimizing a variety of types of structures.  So we need to specify the
 		// type under optimization, in this example Permutation.
-		SimulatedAnnealing<Permutation> sa = SimulatedAnnealing.createInstance(problem, mutation, initializer, hc);
+		SimulatedAnnealing<Permutation> sa = new SimulatedAnnealing<Permutation>(problem, mutation, initializer, hc);
 		
 		// Run the simulated annealer.  The parameter of 1000 means to run
 		// 1000 iterations (i.e., an iteration involves generating a random neighbor with
