### 2020-08-15 #2448 @cicirello
@galessiorob I proofread the latest version, and it looks fine. Thank you for your time and expertise overseeing the review.
